ZFile: align uncompressed entries
Make sure all uncompressed entries are aligned. We used to do the
opposite, i.e. ensure that aligned entries are not compressed. This
caused `zipalign -c` to fail, since the decision about what gets
compressed and what not is usually done by aapt.

For now we'll just use the same logic for choosing which files should be
compressed or not when added directly (not from another ZIP). There's
also a new configuration flag to control what happens to native
libraries (for now they are always packaged as before).

Bug: http://b.android.com/212591
Change-Id: I1cf3b35f0bfc0924bb6061e4f0e4b4753baa7e70