Replace custom stream copy code
I created a new AVD, and the IDE was hanging for 30+ seconds. When
I grabbed a thread dump it showed that it was stuck in this inner loop:

     AvdManager#copyImageFile
        ...
        while ((count = fis.read(buffer)) != -1) {
            fos.write(buffer, 0, count);
        }

I'm not sure whether the streams we're obtaining here are properly
wrapping the underlying file streams with buffers -- but in any case
we shouldn't have custom copying like this since the FileOp supports
file copying directly.

Change-Id: I9defba7f2884c02e3d82b8b0518c566964e584e3