Add Linux bazel scripts to the bazel directory.
The python file is a wrapper intended to be called by any platform and
then it will call the appropriate bazel executable. It attempts to
directly call the checked in version, which should exist in the same
directory. If it does not it will fall back to the system bazel, with a
warning message.

Change-Id: Ic29e5830ecd958534a1f6797490b2c7ee2ddeaf5