DevSdkUpdater now allows specifying package filters
This change obsoletes the need to rely on a synchronized .gitignore
file under the various SDK folders. Instead, a package can be
appended with one or more filters, e.g. changing

"platform-tools" to "platform-tools:{adb*,systrace/**}"

(which would keep adb, or adb.exe on windows, and all contents under
the systrace folder).

Change-Id: Ic2c78ec804a937398fe297849138c07ba1f1744d