Adds support for removing rules.
This CL changes the modifier to delete all rules that have not been
touched in this pass. This means that even though the modifier only
changes "new" rules, if a rule is not "visited" then it will
be removed from the output.

This change applies this new feature to the elimination of "anonymous"
libraries from build files. It will now directly add the imported jar
to the rule.

Change-Id: Icde21a1dc29e08a1b4955c7eba1db94e3122b7ef