Fix spurious warning with manifest placeholder substitution.
When merging libraries' manifests into the application's manifest, the
manifest merger performs several iterations of placeholder substitution:
The first iteration subtitutes the library mafifests' placeholders
before loading values from the application's manifest, whereas the last
iteration does this after loading values from the application's
manifest.

If any issue occurs during the first iteration, we should display it as
an INFO and not ERROR message, since these issues might be resolve
during the last iteration. The current code does this by checking the
current merge type: If it's a LIBRARY type, then it displays the issue
as INFO; otherwise, it displays the issue as ERROR.

This bug is caused by an incorrect merge type passed to the placeholder
substitution method. In the above scenario, during the first iteration,
the merge type should be LIBRARY, even though the purpose of the entire
iterations is to merge into the application's manifest.

Issue: http://b.android.com/225114
Change-Id: I5ec7f3c4c0815b64c034675e9cc7370eee77ed66