Add lint baseline support
This feature allows you to save a snapshot of your project's current
set of warnings, and then use that as a baseline for future inspection
runs where only new issues are reported. This lets you start using lint
to fail the build without having to go back and address all existing
issues first.

Also includes a fix for the following bugs (unrelated to baselines, but
I was touching code in the same area) :
216762: Misleading Lint report filename
78580: Suppress successful lint output in Gradle

Change-Id: I696ee09e77ce26b94eea15165c681b98c0623d8c