Make static variables work with multiple class loaders.
This class addresses the limitation of static variables in the presence
of multiple class loaders.

Since a static variable is associated with a class, it is not uncommon
to assume that a public static variable (or the returned value of a
public static method) is unique and accessible across the entire JVM.
This assumption is correct as long as the defining class of the static
variable is loaded by only one class loader.

However, if the defining class is loaded by different class loaders,
the JVM considers them as different classes, and this assumption will
break (i.e., a static variable is not unique to the JVM but only to the
class loader within the JVM).

The introduced JvmWideVariable class allows creating a JVM-wide variable
that is accessible across the entire JVM despite that its defining class
might be loaded by different class loaders.

Test: JvmWideVariableTest
Change-Id: I3d2d91f6202813ef61236dd4cdefaba43296743e