changed from lambda (method reference) to actual interface.
The aapt2 abstraction layer changed the type to use a super type which
did not include the proper thread shutdown methods. Instead, it was
relying an object cast to make the call or not. Since we passed a method
reference to the function which is a lambda, the instanceof would always
return false, yielding the aapt processes to not be killed.

Bug: b.android.com/229745
Test: adapted existing tests.
Change-Id: I6df5fe6acc61179fda90a2b297858a8975884ef4