Fix regression problem with Instant Run performance tests.
Two of the InstantRun benchmarks seem to have been performing very
poorly lately.

The reason is that in an InstantRun measurement, we (accidentally) run
the 2 builds differently: The first build is run with
project.executor(), which does not disable the build cache; The second
build is run with executor(), which disables the build cache.

It seems that we should always run a build with executor() instead of
project.executor().

This fix replaces the 2 remanining places in performance tests that are
still using project.executor(). It should fix the performance regression
that we see with InstantRun.

Test: Re-measure IR perf tests, build times should lower to normal
Change-Id: Ia9c8bd991747d598f5bd404a1e8894ec4e2a4157