More granular support for library jars.
Previously the internal publishing used the same
jars that are put in the AARs. However this is not
efficient since the main jars contains both resources
and classes. Any change to this jar would trigger the
consumers to attempt to both compile java code and
process jar resources which can be both time consuming.

In particular, in Android, it's rare to touch java
resources, yet the build must support it, and with large
projects and large dependency graph this can be time consuming.

This change introduces a new Transform that process the jars
for the internal publishing only. The old transform is still
there when building the AAR.

In particular this new transform creates 2 separate "main"
jar, one for the main classes and one for the resources.
In addition, it creates local jars like the old transforms,
containing only classes (they are stripped of their resources).

The new transform is also somewhat incremental to only generate
the jars that need to be recreated.

Test: new unit tests for new transform.

Change-Id: I2da13dd23457115a540d3e6db6d5573255c35201