instant run build info: tweak definition of a patch build
When installing split apks, the IDE uses the -p option (patch install)
whenever the results indicate that it is not a full build. This causes
issues during installation if something in the manifest has changed,
and gradle did indeed generate all the splits, but the IDE did not
know that all the splits were generated.

This CL tweaks the definition slightly: The IDE will now use the -p
option if InstantRunBuildInfo.isPatchBuild is true. And a build is
considered a patch build only if it doesn't include all the splits.

There is a minor hack though - the definition of what constitutes "all
the splits" is duplicated across gradle and InstantRunBuildInfo. The
existing bug will be kept alive until that duplication is removed.

Fixes https://code.google.com/p/android/issues/detail?id=231563
Test: unit test updated

Change-Id: I98ac032ad126cec9a39269cd399e09b5b0bdb5c3
(cherry picked from commit 49c6e42b099078756911a24a15c96a3dc617264e)