Detect and fail if different plugin versions are loaded.
If different plugin versions are applied to multiple subprojects in a
build (or multiple projects in a composite build), we will fail the
build.

It is still possible to use different plugin versions in different
builds (e.g., the user can update the plugin to a newer version while
reusing the same Gradle daemon).

This commit introduces the BuildSessionSingleton object that shares the
life cycle of a build. This singleton is created when the plugin is
first applied to a project and released when the build is finished. In
this commit, we use it to detect plugin version conflict, but we will
also use it for other purposes later on.

Test: BuildSessionHelperTest, JvmWideVariableTest; Build a project with
different plugin versions applied to subprojects and check that
building/syncing fails

Change-Id: Ib2fa23c07a255cb5b00b4a4c253905a92d669ace