Fixing the encoding of const-wide 64bit constants
The ART team (thanks Alex!) ran the runtime test suite with dexter -x full_rewrite
and found a problem with the encoding of ceirtain 64bit constants (b/37240589).

This is the fix for the bug, plus another minor change requested by the ART team:
don't print "Running experiment ..." unless -v is on the command line.

Test: bazel test :dexter_tests

Change-Id: I087d0ce8ca43673866af03edac58b8c78cabab15