Allow MergeOutputWriter to delete a non-empty directory.
Currently, if MergeOutputWriter.remove(String) is called on a non-empty
directory, the deletion will fail because it is using File.delete()
which does not accept a non-empty directory. We should instead use
FileUtils.deletePath(File), which accepts a non-empty directory.

This is one of the two issues found in http://b.android.com/356724. The
other issue is being investigated separately.

Test: MergeOutputWritersTest
Bug: http://b.android.com/356724
Change-Id: Ie47adc777056f2f5fc85c06839d906c0b131fb4d