Profilers with JVMTI
This is the first step that brings the jvmti-based profilers to be feature
parity with what we have for the non-jvmti ones.

supportlib and perfa share some classes in the common directory.
These classes are unmodified and no behavior is changed. When a class needs
to be forked, we will copy it like we have ProfilerService now.

Test: Not yet.
Change-Id: I19d3de81d7ec9316f588cb769c68f5994e795d0a