Fix annotations extraction on Windows
Annotation extraction was not working correctly on Windows, as
discovered by our integration test.

The root problem was that the code which tries to compute the
source roots (from individual source files in the project) was
calling into gradle's file visitor "getRelativePath" method and
comparing that path to the suffix of the path -- but
getRelativePath uses Unix file separators instead of native
path separators. This meant that it never found a source path
match, so the type resolver couldn't find the source files, and
consequently resolving fields failed.

Test: This was already covered by a test which was failing (and
therefore disabled) on Windows. This CL adds the test back for
Windows.

Change-Id: If0a889bc66dabf86d89067afa0bbaec58f73037f