Update templates for new Feature Module creation
This CL makes two updates to the templates for new Feature
Modules.

Firstly it doesn't an application element other than in the
base feature. This avoids duplicate declaration amongst
multiple feature modules.

Secondly it automatically generates a dependency on the base
feature from a new feature module and adds a feature
dependency to the base module on the new feature module.
This means that newly added modules will compile and work
without the user having to manually add the dependencies.

Fixes: http://b/37805289

Test: NewInstantAppModuleTest and TemplateTest

Change-Id: Iee095581a1618e30390ca8302c1b0a81e5ef3d52