Always pass min sdk to dx
New dx features (non-abstract interface method, and
signature-polymorphic methods) can be used without
java 8 support. This CL injects the min sdk version
that is calculated based on the DSL and manifest. Also,
if min sdk version is higher or eq than 24, and a version
of dx is used that does not support min sdk version (it
requires build tools 26.0.0-rc2), a warning is logged.

Test: DxFeaturesTest
Change-Id: I0f1e5599894583ea06132f83bb190a557e66e2eb