Remove min sdk from the dexing mode
Dexing type already specifies everything
we need to decide how to merge dex files.
There is no need to correlate min sdk
with the dexing type, once the decision on
the dexing type is made. Min sdk version is
injected separately for all dexing purposes.

We will change legacy to native multidex if
the deployment device supports it.

Test: refactoring, existing
Change-Id: I45724513065f64862630b52c841e767982d8c05b