Fix Kotlin source position lookup
The code to compute the right location offsets also checks to see if
the location is contained in the right file; if not, it will ignore
the warning for in-editor scanning.  This was broken for Kotlin, where
the UAST bridge places name identifiers in a dummy PsiFile that
doesn't match the source file. This made several lint checks not show
results while editing in the IDE.

This fixes
62373283: "Missing Parcelable CREATOR field" is not reported for Kotlin sources
62374648: "Missing Super Call" is not reported for Kotlin sources

Test: Manual; we don't have Kotlin supported from the test
infrastructure yet.

Change-Id: I70a2f2350bf53248995d1f42b14967cb6f9220fa