Don't preverify output for fro main dex list
When running proguard to calculate shrunk jar,
we were also recalcualting stack frames. For that,
we need the entire classpath, so adding shrunk android.jar
is not enough.

However, becuase we always dex the proguard output, there
is no need to recalculate stack frames in the first place.
This CL sets -dontpreverify flag for proguard config in this
particular case.

Test: manual
BUG=62778354

Change-Id: I980c1ae2eddaa06d571f0d060da35743e0eb5283