Disable version check when applying plugin to a project.
Recently, we introduced a precondition check (Change-Id:
Ib2fa23c07a255cb5b00b4a4c253905a92d669ace) that throws an exception if
the plugin is applied twice to the same project within a build. (We know
the boundary of a build by observing buildFinished() events being fired.)

However, with configure-on-demand enabled, when syncing a project, we
don't see buildFinished() gets fired, which causes us to throw an
exception.

In this commit, we temporarily bypass this check until we know the
proper way to handle the case.

Test: Manual (Open a brand-new project, set
org.gradle.configureondemand=true in the gradle.properties file. Before
this change, syncing would fail; after this change, syncing will
succeed). I will create an integration test in a subsequent CL.

Bug: 63329112, 62100695
Change-Id: Ie465c12c1a3b307cd843cc210afd7b08c019b5c1