Fix @Input, @Output annotations in PackageAndroidArtifact.
Recently we replaced the @OutputDirectory annotation in
PackageAndroidArtifact with @OutputFiles to fix bug 38481325
(commit e8d07dc9d58f5c0a0caafa039125ade9f5874177).

This was not an ideal fix, it required a hack of setting @OutputFiles
also as @Input (for reasons explained in the code comment), and it also
makes the task not cacheable.

This commit reverts back to using @OutputDirectory and sets injected
ABIs and densities as @Input to properly address the above issues.

Test: Existing AbiChangeTest
Bug: 38481325
Change-Id: Id1c93e9489ed4581a1693121a1292efb16a7172f