Make specifying old build tools version a warning.
It used to be an error, but now it is just automatically upgraded to use
the default version instead, with a sync warning suggesting to remove
the explicit build tools version from the project.

This should make upgrading Android Gradle Plugin version easier, as
there is one fewer version number to update.

Test: PluginDslTest, BuildToolsTest.
Change-Id: Idadb4b69036a9fcb482c9551c903a24a15641bd4