Re-enable plugin version consistency check.
We want to enforce that, within a build, different projects must apply
the same version of the plugin. This is required so that our build
cache, variant matching, and probably a few other aspects of the plugin
can work properly.

Here, a build refers to the entire Gradle build, which includes included
builds in the case of composite builds. Note that the Gradle daemon
never executes two builds at the same time.

The scope of the check is per build. It is okay for a project to apply
different plugin versions in *different* builds.

The check was previously disabled due to a bug with Gradle's
configure-on-demand feature (http://b/63178989). Now that Gradle has
fixed that bug, we can safely re-enable it.

This commit also:
 - Simplifies BuildSession singleton by letting it live across builds.
   This avoid potential issues when the variables are not reset properly
   at the end of a bulid.
 - Fixes and re-enables the previously-disabled tests.

Test: BuildSessionImplTest, PluginInitializerTest
Bug: http://b/63329112, http://b/63178989, http://b/62878541,
     http://b/64579026
Change-Id: I8a04a567f16ec172bddd7e0f90069833f4907da7
(cherry picked from commit 18c27f5ec75d84c41bc54d1becd14fa7c20a4d82)