Simpleperf should be run from shell context
Simpleperf recently introduced an option (--app) when recording method
traces. This allows us to run simpleperf from shell context, instead of
directly using run-as.

The advantages of that is we avoid the back and forth of switching
between app's and shell context when recording, generating trace files,
and moving the trace to a directory that we can access via adb pull.

This change also adds the --exit-with-parent option to simpleperf record
command, as this is now required to kill simpleperf when its parent
process (in our case, perfd) dies.

As a result of this change, we could remove some steps, such as copying
the trace file file to a temporary file and move it from app directory
to perfd's, and clean up the code.

Test: Manual tests
Bug: 37294023
Change-Id: I508a336f946a89edbda381bc5b77b2b314aac48e