64950482: UI tests failing from "Already disposed" error
This CL makes lint catch and handle some exceptions thrown by the IDE,
such as "Already Disposed" exceptions. In the past we've attempted to
solve this by adding checks for module.isDisposed in strategic places,
but this is a losing proposition: Lint is running on a background
thread in the IDE, and a different thread can close the project at any
time during this analysis -- including between the time we check
isDisposed() and the code right after it that is using the module
data.

Therefore, instead, we catch these exceptions (along with
ProcessCanceledException and IndexNotReadyException, which were
already handled), and gracefully abort lint analysis in this case.

Test: Test included. Plus of course the regular sanity tests listed
in b/64950482 that have been regularly failing before this.

Inspect: OFF
(many enforced false positives)

Change-Id: I7f6d874cded771069e6eba7cb4e5602810cc2a0f