Copy Java resources to a jar file instead of a directory.
In MergeJavaResourcesTransform, when resources inside a jar file are
extracted to a directory, the results may not be expected on Windows if
the file names end with ".", or if there is an uppercase/lowercase
conflict. To work around this issue, we copy these resources to a jar
file.

Bug: 65337573
Test: New MergeJavaResourcesTransformTest
Change-Id: I544724b115e54956d25547e8c09f1fca88cf420b
(cherry picked from commit 0b804415d9cb21e97923aca85fbd46433dd4156a)