Add stream names to transform inputs
In case inputs are added/removed, we would
hit issues in Gradle if incremental build
is triggered. We work around this issue by
passing the sorted stream names as input
to the transform task. In case any of the
streams i.e. inputs is added/removed, this
input property will change, and non-incremental
build will be triggered.

BUG=65573026
Test: JacocoTransformTest

Change-Id: I7e8700fa6397525019816d353db5beffa31c4bab