Disable WordPressAppSmokeTest and WordPressShrinkerTest.
They have been flaky for months but we never figured out the root cause.
We need to temporarily disable them to unblock PSQ.

We may let the tests run only in postsubmit, but that is arguable and
can be done in a later CL.

The only other WordPress-related test that may also be flaky is
NewMultiDexMainDexListTest.checkNewAndOldGiveSameResults[WORDPRESS].
We haven't seen it fail lately so let's keep it running for now.

Bug: 65646366
Test: Disabling tests
Change-Id: Iace61084508c4e20c38088a0f09f3c2af3ab0ad0