Set the feature manifest dependency to the directory
Previously, the MergeManifests task would depend on the output.json,
which is not updated when only the contents of the manifests are
updated. The task now depends on the directory, which includes all
file changes in it.

BUG=67324459
Test: FeatureTest.incremantalManifestBuild

Change-Id: Icd0e44e8e2c37be19bea4579f442f7e111114b20
(cherry picked from commit a2bd32313452a6760b04cb6cdf27ab446e0a5e9f)