More work on DslModelData.
Fix the generation of the sourcesets and
configurations to be as flavor/build types
are created as they are needed during configuration.

This requires changing how we create android Test
variants. Because a single one was created based on
android.testBuildType, however this was not properly
working around source sets.

For instance as soon as the plugin was applied, the debug
and release build types would be created. Their
source sets would be created as well and stored in
the BuildTypeData class. However only debugAndroidTest
source set was created because this was the value of
android.testBuildType. Changing this afterward in the
build.gradle file would not change the BuildTypeData
for release or debug, breaking the relationship between
this extension value and the generated source sets.

There is no easy way to fix this and since we've been
talking about unifying androidTest and unitTest in their
behavior (create matching test variants for all variants),
this is a good time to do this.

Start writing tests for DslModelDataImpl. This requires
quite a few new Fake classes to handle all the
abstraction layer. Also put a few more abstraction layer
in there to avoid having object contain the Project.

Also moved a couple of interface from Builder where they
were not referenced at all to gradle-core.

Test: new tests on DlsModelData. WIP.
Change-Id: I5b1f66f39c9d12ee3f3ad6711373cf9ec90af7b5