Fix src file in android-gradle-build json
Vanilla-CMake returns the source of the file as a relative path and this
could lead to a situation where the IML will end up with a wrong folder
leading to issues as stated in the bug.

Fix: use the source directory information to get the full path

Bug: 67807485

Test: UT's
updated integration test to test for the source file generated in json
file.

android-gradle-build-fixes

Change-Id: Ic9c285771ffd186d7ab30403133d8ab4c9b461b5