Move the code from the plugin into delegate.
There is a main PluginDelegate that will drive the
whole plugin (and all of them), from before evaluation
to the end of afterEvaluation.

There is a secondary delegate (TypedPluginDelegate) that is
provided by the specific plugins (appPlugin, LibPlugin), that
provide plugin-type specific data, factory, etc...
This used to be shared in the plugins and in the specific
VariantFactory.

Test: WIP for now.
Change-Id: I6888df335e1d50ee3d7382bbe2fdfc9d557dc0d8