Factor out the executor configuration in benchmarks.
This is a small change with a larger goal that I'd like feedback on. I
want to shard these tests out so that each combination of scenario +
benchmark + benchmarkMode can run independently.

To achieve the above, I intend to create a Benchmark class that has a
run() method and then construct N of these that can be run
independently, without the looping and switch statements present in the
current code. How does that sound?

Change-Id: I64288c97a06e1a55236cace364c118ef944a7db7
Test: presubmit