Fix for Waze bug on Ubuntu/Mac.
Waze project had a flag like this in Android.mk:

  LOCAL_CPPFLAGS += -DTOM='hello world'

Where the flag had spaces and quotes around the define's value.

Existing POSIX unquoting logic remove the ticks so the user
got a flag like:

   -DTOM=hello world

In the resulting JSON file. Since this results in a command-line
that doesn't compile the user's editing experience was 
broken and very slow (due to failure after failure)

Bug: issuetracker.google.com/69110338

Test: I added a new test to cover Waze's case and I adjusted
the existing parser test to align with the new logic.

ndk-build-quote

Change-Id: I4a2c06444b2a47bb511d0ea5898ca881a7072ce1