Cache eviction step 2: Delete old build cache directories.
The first form of cache eviction is deleting old cache entries of the
current Android Gradle plugin version. This was done in commit
2886c30a2fe11501b37e93592e1f092cf1825d74.

This commit implements the second form of cache eviction: deleting old
cache directories created by *previous* Android Gradle plugin versions.

The cache directory life time (the maximum number of days that a cache
created by an older plugin version can be kept since it was last used)
is currently set to 30 days.

We can allow the users to customize this option later if there is a need
for it.

Bug: 69562033
Test: BuildCacheUtilsTest, AndroidGradlePluginVersionTest

Change-Id: I0210c96cfdfcabfd0176b102e61519c730a5d8ec