Fix task dependency on local file
Local Jar was assumed to be files with no task dependencies, but this may not be
true.  The files may be created by a task.  The transform must be set to depend
on the tasks that generates the jars.

Change-Id: I7f22bbc08bcdfe9ef06b5d3fc2501505bdaf34d4
Fixes: 63076223
Test: Updated LocalJarTest
(cherry picked from commit f279ee2973de18272501723b8aa774bd4ca7e2c0)