Fix NoClassDefFoundError when using Gradle 4.3 or earlier.
ModelBuilder was recently changed to extend a Gradle class that exists
only in Gradle 4.4 and later. If Gradle 4.3 or earlier is used, the
plugin will fail with a NoClassDefFoundError, which can be confusing to
users.

This commit moves the call to ModelBuilder to after the Gradle version
check, so that if an old Gradle vesion is used, the plugin will display
the familiar message asking the user to update their Gradle version.

Bug: 73383831
Test: Manually tested on a HelloWorld project with different Gradle
      versions

Change-Id: I11f9b012af33f005f72ca3eb525eeda57004cc98