Recognise the RequiresApi annotation.
Instant run currently checks if a class is targeting a newer platform by
looking for the TargetApi annotation. If that annotation is not present,
it assumes that the class isn't targeting a newer platform and attempts
to instrument it.

We have a new annotation: RequiresApi. This aims to replace TargetApi by
serving the same purpose. Instant run needs to also honour this
annotation in the same way it honours TargetApi.

Bug: 72811718
Test: added a new test to cover this case in IncrementalVisitorTest
Change-Id: I056bb497175d67877cafed385332f754110b7202