Removes dependencies from ProfilerService
In order to avoid every object knowing every
other object, it is better to limit the responsibility
of each object, and reduce the number of "edges".

In preparation for the elimination of all the services,
this CL first focuses on ProfilerService. ProfilerService
will be a thin layer that delegates to the main object
(the Daemon).

This clean up impacted the way objects are referenced,
which in this case it should be with a pointer. Passing
references and storing them creates the confusion of
local variable syntax with pointer semantic that must
be avoided.

Bug: 73538507
Test: refactoring only, existing tests
Change-Id: Ib89851a4883a9e9d3cedb6e3350aca2aeb78bb15