Add Bundle->Apk flow in Gradle.
New task to from bundle to zip of (all) APKs, and then task
to go from this zip to a folder of the APKs to install with the
help of a device configuration.

Also add new information in the models:
Sync Model: add task names to generate bundle and to generate
apks from bundle.

Post Build Model: new model for post bundle build.

Test: Update DynamicAppTest
Bug: 76022806

Change-Id: I729d0c783d895f20113ffa41d6a9daaabd155c68