Wrap every method visitor to handle JSR/RET instructions
This CL adds JSRInlinerAdapter to every method visitor,
which fixes issues with JSR/RET instructions and stack
frames computations.

It used to be the case that only instrumented methods would
use JSRInlinerAdapter, and that would cause issues if jsr/ret
instructions would appear in e.g. <clinit> and other methods
that are not instrumented.

BUG: 73576747
Test: JsrRetUser.j
Change-Id: Ib0b8e8e5af1687d1855c626fbe51a59262b5613a