Disable configure-on-demand with Gradle 4.6+ in AGP 3.1.2.
Configuration on demand does not work with Android Gradle plugin 3.0.x
and 3.1.x if the user is using Gradle 4.6 or above.

It works with AGP 3.2.x.

This commit will fail the build and ask the user to either disable
configure-on-demand or use a Gradle version less than 4.6. This applies
to only AGP 3.1.2, not AGP 3.2.x.

Bug: 77910727
Test: No automatic test since we don't have Gradle 4.6 in the 3.1-dev
      branch
Manual test: Add org.gradle.configureondemand=true to the
      gradle.properties file and update to Gradle 4.6 in the
      gradle-wrapper.properties file, the build should fail. Either
      removing org.gradle.configureondemand=true, or setting
      org.gradle.configureondemand=false, or using Gradle version
      4.4/4.5 should make build/sync succeed.

Change-Id: Ia7d9756a8d8ff83a3a08355dce8946f38c09276e