Performance fix: don't compute inactive source providers from gradle
In the IDE, only one variant is active -- but we don't want to
show a resource as unused if it's used only from some other
variant -- so the lint check peeks at other variants and includes
their source folders.`

In Gradle we already run lint multiple times on each active
variant; we don't want the lint detector go and add in its own
source providers for other variants.

This was apparently a huge performance problems on projects
with a lot of variants; see issue 71868725.

Bug: 71868725 (partial)
Test: Covered by existing tests
Change-Id: I04f75fa15564e53d3e47537e90b6f1b30280000e