GradleProfileLogger collects all task data before commit
Previously, each call to Benchmark.log() was creating a new Metric
object, adding the data, then committing the result.
ProfilerToBenchmarkAdapter now keeps track of all the Metrics it has
seen, each of which contains all of the data from the Benchmark so far.
At the end of the scenario, the new commit() method on the Adapter class
commits all of these metrics at once, This way, new data doesn't
overwrite old data.

Bug: N/A
Test: Manually verified outputs
Change-Id: I21f4470d19a06adc01b53ebb5210ac50206eacfe