Fix DataBindingIncrementalTest failing on tools build bot.
The test passed PSQ when the related CL (commit
ce66ca0c6410b3da89528ea59e44168669118d31) was merged. However, it
is currently failing on tools_studio-3.2-dev.

The cause of this issue is that an error message may be printed out on
stdout on some of the bots and stderr on the others, resulting in failed
assertions in the latter group.

To fix that, this CL makes sure both stdout and stderr are accounted for
in the test assertion.

Bug: N/A
Test: Fixing it
Change-Id: I8bad4c9974ec0bb41b607e533b2d5831100145c8