Fix DataBindingIncrementalTest failing on tools build bot.
The test passed PSQ when the related CL (commit
c12c20397b9e86e3edfc9b0578f102bdb7944b5d) was merged. However, it
is currently failing on tools_studio-3.2-dev.

The cause of this issue is that an error message may be printed out on
stdout on some of the bots and stderr on the others, resulting in failed
assertions in the latter group.

To fix that, this CL makes sure both stdout and stderr are accounted for
in the test assertion.

Bug: N/A
Test: Fixing it
Change-Id: I8bad4c9974ec0bb41b607e533b2d5831100145c8