Base infrastructure for lazy tasks.
New factory methods. Duplicates of the existing one
but provide lazy versions.

Refactor TaskCreationAction for direct and lazy versions
as well as config-only interfaces.
This requires changing how we handle preConfigure as we
are not guaranteed to call it both before the Action
itself and after register(). Therefore the method
now only receives taskName and is called via a custom
Action that wraps LazyTaskCreationAction and handle
the lifecylce in case the action is called during
register() because of callbacks like withType or whenAdded
on TaskContainer.

Move all task references to the TaskContainer for more
clean-up.

Since the manifest merging tasks already used the injection
of BuildArtifact, they directly are created with the lazy
factory methods.

Bug: N/A (internal clean-up)
Change-Id: I4873881afb8dd87ef4148a261c52da4254281740
Test: existing ones.