[AVD Mgr] Correct hw.ramSize in config.ini
When a device's hardware definition changes, we update the
config.ini of AVDs that are based on this hardware.

The update function was extended to indicate how much RAM the
device has, but the amount was written to the file incorrectly
(e.g., as "4 GB" rather than as "4096").

CL introducing the bug: b3341e7a12744e989bb4c2eeefa4b983abe94995

Bug: 112664026
Test: Extended DeviceManagerTest.testGetHardwareProperties
Change-Id: I0c38df3c7da8b648b65b3c3170c30489d4ee348e