Add navigation option for the BasicAcitivty template
This is part of the larger effort to add a navigation option for each
Activity.

This CL allows to add a Navigation Drawer to the Basic Activity using
the navigation option.
To try this, you need to enable the NPW_NAVIGATION_SUPPORT flag
added in ag/4937800 first, then in the New Project Wizard, Proceed as
-> Mobile/Tablet -> Basic Activity (check Include Navigation Controller)
-> Choose Navigation Drawer from the combobox

The default version for gradle plugin version is still configured as
2.3.x, so to get the generated project to work, the version needs to be
modified 3.x version (e.g. 3.2.0-beta03).
Modifying the gradle plugin version is coming in a future CL.

Bug: 113988003
Bug: 114327516
Test: Added in a future CL
Change-Id: I0787b92b01b09ac477c4f2e2f3dceec297e55909