gen*Sources does not need to generate R.jar
If the option to generate R.jar instead of R.java is enabled
then make the generate*Sources task not depend on the resource compilation
task since this is not needed (the IDE does its part by creating virtual R
classes from the resources instead of reading the generated R.jar).

Performance of the large test project for running:
  ./gradlew generateDebugSources generateDebugAndroidTestSources
goes from about 80 seconds down to 5 seconds.
This is a huge gain for post-sync build time.

Bug: N/A
Change-Id: I85349cd0e24bb0c88fee5b084cbcdbeeea94fa60
Test: existing tests