Make GlobalScope available to the BaseExtension
Fold the AndroidBuilder inside the GlobalScope so that
the only way to get it is via the GlobalScope. Before
many objects ended up with both.

Also fold the NdkHandler inside the GlobalScope.
More clean up work will move more things inside the global
scope to avoid having a lot of objects get tons of fields.

Change-Id: I971afc5d5af318ef85434c28162b6db1fb61857c
Test: existing tests.
Bug: 115775518