Add Tabs navigation option to the BasicActivity
This is part of the larger effort to add a navigation option for each
Activity.

This CL allows to add a Tabs to the Basic Activity using
the navigation option.
To try this, you need to enable the NPW_NAVIGATION_SUPPORT flag
added in ag/4937800 first, then in the New Project Wizard, Proceed as
-> Mobile/Tablet -> Basic Activity (check Include Navigation Controller)
-> Choose Tabs from the combobox

Test: Coming in a later CL once this feature is not behind the flag
Test-Strategy: ui.test.groups=PROJECT_SUPPORT,PROJECT_WIZARD,THEME,EDITING,DEFAULT

Bug: 113988003
Bug: 117129763
Change-Id: I5b920f9505aa1287eda6a6afea0b321e53da7175