Parse AndroidManifest.xml to obtain process names.
Modifies the deployer to parse the application manifest to determine
which process names it needs to look for.

Updates the installer to spawn a server for as many agents as there
are running application processes, rather than a hardcoded value of one.

Test uses a prebuilt APK with a specific manifest; should probably
automatically generate that in the future.

Bug: 116740217
Test: ApkTest#testGetApkDetails()
Change-Id: I9484111c9ab710b8b64ca68e594f19063a62c0a1