Fix caching same APK twice corrupts the DB.
The problem with the current DB exceptions is rooted by a bug that
creates duplicated dex file entries in an APK when it is cached twice.

For example, if APK = {dex01, dex02}, calling cache() twice will result
APK = {dex01, dex02, dex01, dex02} when retrieved from the database.

The reason why we don't see this very offen is because we when create
a Java representation of the APK cache, we use a hash map of filenames
as keys. This masked the problem with duplicated entries.

However, as more swap invocation happens, some of the dex files will
get old and eventually deleted. Depending which dexfile gets deleted
we can ended up with a situation where the duplicated dex is deleted.
Because of the indeterministic nature of how the results are returned,
the older file index *could* get returned. Subsequence modification of
that dex file will use an deleted dex and therefore ends with an
foreign key error.

In particular, this CL:
-Output the result is sorted order to avoid indeterminism.
-Add a check for duplicated entries in an APK
-Change the DBs so that caching the same APK twice becomes no-op.
-Add a set of test that would check all combination of DBs from the
 dex API level instead of just the DB level.

Bug: 117440317
Bug: 117332999
Bug: 112378806

Test: DexArchiveTest, SQLiteDexArchiveDatabaseTest

Change-Id: I61b5e31f2321480d418ef8b31d8c1ccb576df534