Implements the new pipeline for "Apply Changes"
This change defines clearly all the steps in the swap and are added
to the Deployer swap. Each step has a clear, immutable, inputs and outputs.
This paves the way for parallel task execution which will be done in
follow up CLs.

In order to have consistent datatypes for input and outputs, this
change unifies the various Dex/Apk representations we had.
OnCacheDexFile, OnHostDexFile, ApkFull, ApkDump, they all become
a simple set of data classes.

The database API is simplified and we only support the SQL version
now.

Test: updated, more to come.
Bug: 117350620, 116481986
Change-Id: Id0e1c1680e0d16af33e65d2c6194d667bcfa363c