Add INTERMEDIARY_BUNDLE artifact and split Bundle signing.
This new artifact is a lightweight intermediary version of our final
bundle artifact, that skips some steps not needed in order to improve
performance of tasks that will use the bundle to produce one or more
APKs.

We now have a new task, that copies the intermediary bundle to it's
final location and also signs it if it's the case (for now, if it's not
in a debuggable variant).

Bug: 118474005
Test: existing
Change-Id: Iac14dc9239b939fee92a3731f9874e9b46c72a14