Allow users to blacklist libraries from Jetifier (1/2).
Jetifier is supposed to work on third-party libraries that do not yet
support AndroidX. However, if a library has already been updated to
support AndroidX, or if it is not using AndroidX/support libraries at
all, then it is not necessary to run Jetifier to convert it, and doing
so may even result in incorrect rewritten binaries.

While we work on a proper fix for this issue, let's unblock users by
allowing them to blacklist libraries that should not be jetified.

Users can now add a comma-separated list of regular expressions to the
following property in the gradle.properties file:
    android.jetifier.blacklist = foo, bar

If a library's absolute path contains a substring that matches one of
the regular expressions, the library won't be jetified.

For example, if the regular expression is  "doNot.*\.jar", then
"/path/to/doNotJetify.jar" won't be jetified.

Bug: 119135578
Test: New test in JetifierTest
Change-Id: I6531fe0faafa45dc2811c5223ba199ccce5fee53
(cherry picked from commit d9dedb56beb93eb881c03bf6317cdde14159abeb)