Add Lint memory leak test
This test attempts to catch memory Leaks by running Lint on a few files
and asserting that there are no strongly reachable instances of
LintCoreProjectEnvironment, PsiWhiteSpaceImpl, etc.

Ideally we could test for memory leaks after running Lint on a very
large and complex project using a Gradle daemon, but that will be
trickier to do. Still, this test is able to successfully detect
the bug b/119833503 fixed by change Ia17206f5. If you revert that
change and run this test, it fails with:
"Detected Lint memory leak; LintCoreProjectEnvironment is reachable"

Test: self
Bug: 119833503
Change-Id: Id4cd06bea4e6ee46c9d945f5347e969f9bf5ed78