Fix casting error in GradlePluginUtils for composite builds.
The GradlePluginUtils class is used to detect plugins that are lower
than a minimum required version.

It used an unsafe cast from ComponentSelector to
ModuleComponentSelector, which will break in the case of composite
builds (where the type could be ProjectComponentSelector).

This CL removes the unsafe cast (the code that needs casting is not
extremely useful).

Bug: 121017797
Test: Existing PluginVersionCheckTest (integration test) and
      new GradlePluginUtilsTest (unit test)

Change-Id: I69e5f80399f7b8345f0a5bb53f39c9cd30dfa682
(cherry picked from commit ccdd7a7daa43fcf40bd92b5ece4525101c5bf404)