Calling Attach Agent From Debugger.
This is the workaround lack of JDI RedefineClasses capability
in Oreo.

The idea is to attach our agent when the debugger is attached to
perform business as usually. We take advantage of the debugger's
ability to force invocation of a function, used when printing
expressions while debugging, to invoke ART's VMDebug.attachAgent
to start our agent.

It is not complex enough for a full design doc but there are enough
details in the comments.

Bug: 121151729
Bug: 120988743
Bug: 124531482
Bug: 126945654

Test: manual debugger swaps

Change-Id: Ifbbf7a83d87d51e1041b3fe656a4a7928f6a0724