added facility to record wokers execution.
The current profiling information did not account for Workers,
it would ignore them all together.

With this CL, Workers are properly linked to the task that spawned
them and task timing information is adequately calculated.
Chrome traces have been updated in order to create the span and
sub spans with Workers.

Bug: N/A
Test: manually so far, coming.
Change-Id: I88f69098fcf2cdf7808d4e910e38a59292681d4b