Add test for incremental @Classpath support CL
The original incremental @Classpath support CL (http://ag/6690610)
caused a large regression in the Base100_Java8_javaResAddLib benchmark.
After investigating, I found that there was a regression because the
existing code has a bug, resulting in a noop for the package application
task.

The bug in the existing code results from the inputs to
MergeJavaResourceTask being annotated with
@get:PathSensitive(PathSensitivity.RELATIVE).
All of the libraries create a java resources res.jar file with the same
relative path via the BundleLibraryJavaRes task. When a new library
source java resource file is added, Gradle's IncrementalTaskInputs
reports the wrong library's res.jar as having changed.

Using the @Classpath annotation fixes this bug because the @Classpath
annotation doesn't consider the input file names or paths at all.

This CL adds a test which fails when the MergeJavaResourceTask inputs
are annotated with @InputFiles instead of @Classpath

This CL also replaces the calls to searchAndReplace() from the previous
CL with project.replaceInFile(), which ensures the edits are reversed.

Bug: 126396809
Fixes: 128858509
Test: JavaResPackagingTest
Change-Id: I2f71248613eb756e0840c164802cfa80aaf66a3a