Relax the regex for dependency resolution handler.
Currently the regular expression that extracts the
dependency name from the depedency failure resolution
message restricts the string only to characters valid
for an artifact name, which leads to incomplete (and
hence hardly actionable) failure message in Studio
in case the user makes a random typo in the dependency
name.

This CL relaxes the regexp to include all non-whitespace
characters.

Test: Added DependencyFailureHandlerTest to verify the intended
dependency extraction result.
Bug: 129500977
Change-Id: Idbbace93af9ac955584eed288a64c4623aa27938