Updates 'perfgate_test' tag to 'perfgate_multi_run'
Perfgate_test is not a good tag name, as it implies the filter simply
means 'run this on perfgate'. That is not the intent. The flag is
supposed to mean 'run this test multiple times for more data samples on
perfgate'. Renamed 'perfgate_test' to 'perfgate_multi_run' to make this
something users can understand.

Bug: NA
Test: PSQ
Change-Id: I533d957114ed0840aeed5e4aa99d1eee71828785