Add support for writing aspects agent violations to log
Currently, the aspects agent will make use of the Assertions class to
throw exceptions whenever it finds a violation that is not whitelisted.
In some cases, that might be too noisy and we might want to use a more
silent approach instead.

This CL introduces AspectsLogger, which will append the violations to a
log file controlled by an environment variable (ASPECTS_AGENT_LOG). This
file might be read later by a single test that can fail depending on the
logs content (i.e. if we logger any violations).

Test: AspectsLoggerTest, StackTraceParserTest
Bug: 130264928
Change-Id: Ife96aa08f8ad00b729661b42c123b1d5441466ec