IRR: Zip parsing to account for uncompressed size
Problem: The current zip parsing only uses the "compressed size"
field when parsing a zip archive. This results in patches that are
bigger than necessary since uncompressed entry report a size of 0.

Solution: Detect when compression is used (compression flag is set
to 0 when an entry is not compressed) and use the correct value.

Bug: 130010334
Test: Unit Test
Change-Id: I943a02b0634d2fd78054ef2ed2b1872f8e31bb57