Fix live allocation startup failure
Perfa initialization retransforms loaded classes and instantiates live
allocaiton tracking. The former may take long in a large project and
delay the latter, which is required for live allocation to start when
starting a session. Since the latter is already put in the
daemon-connected callback, we don't have to worry about it being called
too early and can thus move it before class retransformations.

Also added retries to SendRequestToAgent calls in perfd in case agent
attachment took too long.

Fixes: 131105848
Test: manually profile AndroidTank app (unity)
Change-Id: Ie1a269fb0060c9b0acd46379a4ac3aeb84ece312