Lint: remove SparseArray-over-HashMap preference
Removed because, while SparseArray is more memory efficient and cache
friendly than HashMap, it is also generally slower than a HashMap because
lookups require a binary search. It is also not intended to be appropriate
for maps containing a large number of items. Thus it does not make sense to
unconditionally recommend SparseArray. Find more info at
https://developer.android.com/reference/android/util/SparseArray

Test: JavaPerformanceDetectorTest (updated)
Bug: 129760342
Change-Id: Ie989e1da7509b9bf5cc3a23c09b7300848cd843d