Report all D8 desugaring warnings as info
Output was cluttered with desugaring warnings. While those can be
very useful when running standalone D8, within the build system, it
is mostly noise. The reason is that the build system sets up the
classpath, so all available types are already in the desugaring
classpath for the classes compiled in the project. For classes coming
from Maven, there is not much we can do, so it makes sense not to
report these warnings.

Users are still able to show the warnings with "--info".

BUG: 118842646
Test: manually verified
Change-Id: Ie6a10235b5f4b82b66d13ed0f286a008868e584d