Support class level annotations in the aspects agent
We were only supporting method level annotations in the aspects agent,
meaning we would only apply annotation rules if the method itself was
annotated.

This CL makes it possible to intercept all methods in a class if one or
more of its annotations are in the aspects agent rules. If the method is
already annotated with the class-level annotationm, the agent should
intercept it only once.

Bug: 133300743
Test: InstrumentTest, TransformTest
Change-Id: I3780404090955bb5d9a45e950c809e78732625c6