Added jar artifacts for all variants as input to LintPerVariantTask
LintPerVariantTask will eventually call ModelBuilder which will query
Gradle for jar artifacts for other variants, this change adds the jar
artifacts for all variants as input to the task so Gradle would make
sure that they are resolved before starting the task.

Bug: 132783036
Test: Verified manually. Ran FlavorsTest 500 times successfully while
usually it takes 50 times to produce a timeout.

Change-Id: I8ace375442df0efd8fdd0fc5ccdfd3983acccdfd