Change the aspects agent rules file format to JSON
Convert the rules file to JSON to make it easier to read (since most
people are familiar with JSON) and to allow extending it to specify new
features to the agent (e.g. adding the concept of groups to annotations
within the same scope). The new format is the following:

{
  "methods": [
    {
      "name": "java.nio.file.Files.walkFileTree",
      "aspect": "#warn1"
    },
    {
      "name": "java.net.URL.openConnection",
      "aspect": "com.android.tools.checker.AspectsLogger#logIfEdt"
    }
  ],
  "annotations": [
    {
      "name": "com.android.annotations.concurrency.UiThread",
      "aspect": "com.android.tools.checker.AspectsLogger#logIfNotEdt",
      "group": "threading"
    },
    {
      "name": "com.annotations.custom.MyAnnotation",
      "aspect": "#assertIsNotEdt"
    }
  ]
}

README.md was update accordingly. This CL took the opportunity to also
include details of the baseline and how to specify it.

Bug: 134568935
Test: RulesFileTest
Change-Id: I7e1317eebb91bf9c815bb618c57715a8861eb3fc