Separately sends BatchAllocationContexts from allocation and jni ref events
In preparation of the new unified data pipline, this CL eliminates the
unbatching of allocation events in the studio data poller/datastore layer.
Instead, the batched allocation events and contexts are sent as-is to the
Studio UI, which would handle the unbatching.

A number of issues were addressed that was causing flakiness in tests:
- Ensure that the tests wait until the tracking initialization stage
has been completed.
- Switch from calling System.gc() to Runtime.getRuntime().gc(). The first
takes longer and less likely to trigger an actual GC event, as it requires
finalization to run first. This better guarantee the test objects are freed
and generate the deallocation events the tests look for.
- Fix a timing issue where JNI ref events for an object can be cached and
queried before the object's allocation event.

Bug: 121392346
Bug: 80501379
Test: Existing

Change-Id: Id2b76d445e8f6d62aceaca1dc0ac23f2bbcebba9