Support source filtering in coverage reports
Extend coverage_report to allow source path prefix includes and excludes
to enable custom filtering. This enables precise intratarget reports.

Note that these filters are in addition to rather than instead of the
master filter. If something is dropped by the master filter, then a
custom include here will not recover it.

Test: N/A
Bug: N/A
Change-Id: I1d4d52ce09dbad40415cf43559eee3f94cf74132