Stop deleting __content__.json file
Previously, we would delete the __content__.json file for an
IntermediateStream when calling IntermediateStream::asOutput for
non-incremental transforms. We deleted the __content__.json to avoid
errors thrown when a transform's scopes changed between builds (e.g.,
see https://issuetracker.google.com/79090176).

In some cases, deleting the __content__.json can confuse downstream
incremental transforms, specifically if the downstream transforms are
relying on the indices of the SubStreams to remain the same
between builds. In the repro projects for these bugs, the
non-incremental transforms were producing the same number of
SubStreams, with 1 removed and 1 added, and many of the SubStream
indices reassigned. The downstream incremental transforms saw all the
file statuses as CHANGED, instead of any being ADDED or REMOVED, which
eventually led to an error.

This CL stops deleting the __content__.json file and instead ignores the
"Unexpected scopes" error when calling IntermediateStream::asOutput.
This fixes the issue because now a new SubStream will be assigned
a new index, so the downstream incremental transform will see the
desired ADDED status.

Ignoring the "Unexpected scopes" error when calling
IntermediateStream::asOutput doesn't interfere with its being thrown
when desired, as demonstrated by
TransformTaskTest.streamWithTooManyScopes()

Bug: 135967294, 135967286
Test: MinifyFeaturesTest,
  TransformTaskTest.oldContentJsonWithExtraScopes(),
  existing TransformTaskTest.streamWithTooManyScopes(),
  manually tested repro projects
Change-Id: I8e56dd28bb1e43e4584be5d38b83541d4d44fc75