[Aapt Kotlin] Added pseudolocalization
Pseudolocalization primarily used to test Apps by "localizing" the
translatable portions of string resources under "pseudolocales" en-rXA
and ar-rXB. One, which localizes the text of the string and the other
which reverses the direction the string is read, like a right-to-left
language.

The code is broken into two parts, first the Pseudolocalizer which
performs the localization upon each of the strings contents, and the
other is the PseudolocaleGenerator which parses the ResourceTable for
resources to localize, and generates new resources from them.

Bug: 132800341
Test: added tests for both the Pseudolocalizer and
PseudolocaleGenerator.

Change-Id: I6187e2c00f964a78dced1fa8d57f3ce98c627710