Fix inputs that are incorrectly ignored by Gradle.
Due to this Gradle bug: https://github.com/gradle/gradle/issues/10186,
`fun isSomething(): Property<Boolean>` can't be annotated with @Input
(it is ignored and logged as a warning by the task validator).

This CL changes the properties' names so that Kotlin will compile them
into `fun getSomething(): Property<Boolean>` which is accepted as
@Input.

Also fixes a method in ExtractAnnotations where the @Input method does
not start with `get` or `is` and will be ignored by Gradle.

Bug: 135900510
Fixes: 138995868
Test: ValidateTaskPropertiesTest
Change-Id: I1c4f1e8f581639842ee3ec445697aa8372d79c6c