Invalidate local bazel cache for bots uploading results using ATP
Currently only qa_{sanity,fast,unreliable}_linux do this.
The potential issue is that after deleting or renaming a target, the
obsolete test result XML sticks around inside testlogs and would get
pulled into the upload. Bazel clean invalidates the local cache, which
ensures any test result XMLs in testlogs are either pulled from the
remote cache or generated in the most recent run. Using --async does
most of the cleanup in the background, avoiding impacts on script
latency, and it only returns once it is safe to invoke bazel again,
which prevents correctness issues. The runtime difference is minimal
(~100s for sanity).

Test: N/A
Bug: N/A
Change-Id: I282e4f67da4723001733cf6d2557330c0808fa92