Generify ArtifactType
In order to provide type safe APIs in the Variant API like :
fun <F : FileSystemLocation> register(type: ArtifactType<T>, provider: Provider<T>)
it is necessary to add a generic type to ArtifactType indicating
if a particular type is a RegularFile or a Directory.

The current implementation relying on enum classes could not
be extended since enum do not support generic types. The
new ArtifactType<> is now a sealed class with single objects
representing each artifact.

The kotlin code is basically unchanged from using the enum
values to the sealed classes singletons, however, in Java,
INSTANCE had to be added. I could have worked around the issue
by introducing another level of indirection but since we are
moving our code to Kotlin, I felt it was uncessary.

BuildableArtifactHolder has been upgraded to provide type
safe APIs but is otherwise mostly unchanged. In upcoming CLs,
the class will be retrofitted with the new Variant API
implementation. The new Variant API interface is provided as
a reference at this point (Operations.kt) but will be implemented
and enhanced in subsequent CL.

Moved all uses of ArtifatType and subclasses to use the new
generic type, fixed a few misuse of APIs in the process.

Test: Adapted existing.
Bug: 139752844

Change-Id: I1ff77ee3c10c09e3e9b37fce61a152ddf8dad3bb