Reduce the number of targets tested on Macs
This is a temporary (1-2 days) hack to restrict the number
of targets built on a Mac. Many targets seem to be timing out on Macs,
and we haven't had a green build in over a week. Each build seems
to time out after close to 2 hours. Automatic retries (b/142000342)
and reduced Mac capacity (b/141391286) exacerbate the issue.

This CL is an attempt to start from a clean slate, and add targets
progressively.

Bug: 141646507
Test: this
Change-Id: I21b6316fda1d3ed9a872ccb7f2db7942670fea90