UnifiedDiffTest: trim strings before comparing (windows only)
UnifiedDiffTest fails on Windows often. When it fails, the results that
it obtains have \r characters in them, while the golden output doesn't.
It is not clear to me whether the diff should include \r characters or
not, but in the meantime, rather than @Ignoring this test, this CL
changes the comparison in the test to ignore whitespaces.

Bug: 142972246
Test: updated
Change-Id: I4ff4252bf9a45e000985748c227dca75fe846770