Zipflinger: Use default version made by
Zipflinger uses Unix as default OS for "made-by" field in the
Central Directory Record entries. This is a problem since it
sets the external attribute to 0. As a result, tools such as
command-line unzip generates files with permission "----------".

Solution: Use the default (0=DOS) value for "made-by" field which
trigger unzipping tools to default permission to   "-rw-r--r--".

Fixes: 141861587
Test: Unit Test

Change-Id: I0b1f3fb4e0144fda826f87932c589e819ccd0626