Remove stale flag "profiler.cpu.simpleperf.host"; add missing test
Replaced SimpleperfManagerTest#StopProfilingFailToConvertProto
by SimpleperfManagerTest#StopProfilingFailToCopyRawFile
because with the flag being removed, the former test is no
longer applicable. Add the latter as it tests current product
behavior (symbolizing trace on host).

In order to test SimpleperfManager, add FileSystem as a
parameter to its constructor. Unify the return value of
DiskFileSystem::MoveFile() and
MemoryFileSystem::MoveFile(). Document the behavior
in the FileSystem abstract class.

Bug: 144847667
Test: updated; see commit message
Change-Id: I67b4bc09f990f42605fdc5b38d83a16e2b9b8a9a