switch applicationId to use Property<>
Added a new Property<String> on VariantProperties to represent the
applicationId.
Rewired all use of the old API to now use the new property from
the VariantProperties isntance.

Things to do: Stop having the applicationId transfered through sync
as it will not work when external users will use Task to provide a
value. Instead, we should use the post build sync (or its replacement).

Test: Exsiting.
Bug: 139752844
Change-Id: I5f444b55ef2bcd3c2bccca6c076f493cc70a6095