Remove some of the state from workers factory
This CL removes static state used by the
workers factory. Instead, tests can create
instances of WorkerExecutorFacade that are using
the direct executor. Also, as ForkJoinPool.commonPool() was always
used, this parameter is omitted from the initialization
in order to simplify things.

Users of WorkerExecutorFacade now need to specify if they are
using Gradle workers, and that is accomplished by injecting the
flag value into all tasks. In this way, state maintained by the
factory is smaller.

In the next CL, aapt pool will be moved to a Gradle BuildService,
making the factory truly stateless.

BUG: 145129387
Test: updating them
Change-Id: I6f5d4122fe44f74eed07cb41d6aa1f85d87a57bc