Introduce new transport.iml IntelliJ project
For legacy reasons, profilers.iml has been growing to include things
outside the scope of profilers (since the transport API was refactored
from profilers). However, before moving forward with even more features,
we're going to take a moment to split out the non-profiler modules out
of profilers.iml.

To start, this project only contains "fakeandroid" in it for now. In a
followup CL, we'll add a test framework, extracting just the core bits
out of perf-test.

We also updated "fakeandroid" to use "Guava" and "JUnit4" project-level
libraries, which can be shared by future modules added to this project.

Bug: N/A
Test: N/A
Change-Id: I1df58bbc43089178e6ab830a09f26692a9c36ad6