Agent Support for Static Variable Re-Init
This is the bare MVP that support adding an int which
matches up to exisiting tests in StaticVarInitTest.

I tried to make this small enough to review yet functional
enought support the test. Error checking will be added in next CL.

This also does not have D8 support yet. The test, like
the rest of FakeAndroid test, are invoke by manually creating
a request PB instead of using our tools to generate it.

Bug: 142263860
Test: Exisitng StaticVarInitTest
Change-Id: I30799c15c2dc2f15f1431a721bbf5b1cfb5243c0