Add BaseExtesion.buildFeatures
Because buildFeatures is specific to each plugin type,
the method receiving a config action is not present on
BaseExtension as this cannot compile with the more specific
methods in each plugin's extenion.
This makes code (kts or compiled plugins) using BasExtension
unable to change Buildfeatures flags.

This change adds a getter, returning the base BuildFeatures
class on BaseExtension to help with this.

Fixes: 147207696
Test: N/A
Change-Id: Ic1e3e2d15cdeb9852fef476d15417c6053ceab4a