Support full width dollar signs in lint tests to avoid templates
Using "$" in test files is annoying from Kotlin; it requires
a substitution of ${"$"} which makes the source code really
hard to read when you have $-heavy code.

This CL makes it possible to use a fullwidth dollar unicode
character instead, ＄, which does not have to be required and
looks similar to the $ sign. Lint will treat these
interchangeably.

There's also a new (internal only) check in lint sources for
${"$"} and ${'$'} and suggesting these be replaced with the
fullwidth dollar character, ＄.

Test: Unit tests updated
Bug: N/A
Change-Id: I5aaf620d30f7127a3f5aaa79ef3e040bdf44ec3a