Proper handling for appId and debuggable
debuggable is now available ona strict subset of previous
interfaces, and therefore requires the introduction of new
interfaces that are split from the normal hierarchy.

The first such interface is ApkCreationConfig that
is the base for all APK variants/components (app,
dynamic feature, separate test modules, and androidTest
component for all modules).

Moving some TaskCreationAction to this interface
has a cascading effects. Ultimately we need to move
TaskManager to use all these new interfaces, but
this is a very large change, as it will also
require changing how to handle tasks for different
types of plugins, and move logic away from just
using VariantType and instead use type hierarchy.

In the meantime, we use a few safe casts in a few
places.

applicationId is now handled separately in each
XPropertiesImpl class as this is initialized
differently depending on the variant type.
This moves the logic from the Task configuration
to the components where it belongs.

Bug: N/A
Test: existing tests
Change-Id: Idc57dd86465cd3bde107a60969f58c27854d9a6e