Split up connected tests into individual bazel targets
And reduce timeout to 15 mins.

Splitting up the connected tests increases the total machine time to
run all the tests by about 20% because there's some shared overhead when
they're combined, but it reduces the wall time when running the tests in
parallel, and it allows us to reduce the timeout to 15 mins.

Bug: 148626301
Test: these
Change-Id: I350598b60bcee1380f5857d67d769b3a8d6d5589