120255046: Handle more version check methods
In some cases, resolving a method call will fail. However,
we can sometimes tell that a called method is a version check
based on the name of the method; this CL adds this as a
fallback option.

Test: Unit test included
Fixes: 120255046
Change-Id: I2c186a5b3a4f83d5dcc48a1db95fe99a77c06fb5
(cherry picked from commit 01e519150b46dfb05644fad80a3b5db58f8a3957)