Pre-compute build outputs in AndroidArtifactImpl
When getOutputs is called from Studio, the method is executed
through Gradle TAPI invocation. With change I1ff77ee, the ArtifactType
is singleton of Kotlin classes. Gradle tooling API doesn't handle
that well in invocation, and causing the getOutputs method throws
exception.

This CL creates a class field buildOutputs, and compute it once in
constructor, so that no computation will be done through
Gradle tooling API.

Bug: 141163472
Test: NA
Change-Id: I206f1c81d01bbfba24432ea61d28a502456544d9