Create bazel target for MergeNativeDebugMetadataTaskTest
MergeNativeDebugMetadataTaskTest wasn't compiling in gradle or the IDE,
which was surprising since the bazel tests were passing. The reason for
this is because there was no bazel target for this test.

This CL fixes the test and adds the bazel target.

This CL also changes the test to build the debug variant instead of
the release variant because the test had a flaky failure caused by lint,
similar to Issue 128984763.

Bug: NA
Test: existing
Change-Id: Ia1d1a551afb41b63ae799a331803b867f2982559