Add Multidex support dependency only when building APKs
This CL changes when Multidex support library is added as
a compile and runtime dependency. Now, only when building
APKs this dependency will be added. This means that when
publishing POM/Gradle metadata for AARs that have
multiDexEnabled set to true, this support library will not
be added as dependency.

Bug: 146165298
Test: PublishLegacyMultidexLibTest
Change-Id: I7fee310486b9bd432ab7c8282489d26e48b60acc