Add support for updating overlay on swap fail (R+)
This CL adds support to update the overlay even when swap fails,
as long as the user enables fallback on swap failure mode. This
is done by simply passing in a flag to the optimistic swapper to
commit the install even when the swap fails.

This allows us to update the overlay + restart, which saves us
from doing a full re-Run. The restart is necessary to ensure that
the restarted app matches the updated overlay state.

Fixes: 145929571
Test: TBD
Change-Id: I73c2376c5044a77b152b0e29df49bb2f834484ac