Decouple PERFETTO from ATRACE on perfd
This CL creates a CpuTraceType for PERFETTO and use it inside perfd
instead of using Atrace + boolean in proto to use Perfetto to capture.

This will allow us to track all the way which technology we actually
want to use, Perfetto or Atrace, without having to rely on a combination
of inputs.

Bug: 147099951
Test: existing
Change-Id: I1aad9639579f11632853fab689c6f9434b799f13