Don't enter JVMTI callback if already in a callback
A deadlock issue is discovered in a newer version of R.
To fix it and avoid similar issues in the future, we
conservatively avoid entering a JVMTI callback if the
thread is already in a JVMTI callback.

This CL generalizes commit f85c779ad1be8711d73d8911554157d2b1526de5.

Bug: 151177685
Test: manual
Change-Id: I52ac8d859b44157b6070bb2652ed2fafc6fc52a5