Remove dex output for input jars impacted by desugaring
A jar may need to be desugared even if it does not change.
This is because one jar may impact desugaring of another jar.
For jars that were not changed but had to be reprocessed,
AGP failed to remove their outputs.

Another important piece of information is that output path
for input jar differs depending on if it is retrived from
the Android build cache or not. If there is a cache hit,
output path is "$hash.jar", and if not output path is
"$hash_<number>.jar".

This means that in the following scenario we will end up with
duplicate classes in the output:
- there are 2 external libraries, libA and libB, and desugaring
  of libB depends on libA
- there is a full build, followed by clean, followed by full build
- right now there are 2 outputs "$hashA.jar" and "$hashB.jar"
- libA changes in the next build
- we determine that libB will need to be reprocesed
- we remove "$hashA.jar" but we fail to remove "$hashB.jar"
- after the task finishes output contains "$newHashA_0.jar",
  "$hashB_0.jar", "$hashB.jar"
- dex merging task will fail as classes from libB are duplicates

This CL makes sure that we remove output of jars impacted by
desugaring before we start processing them.

Bug: 152996306
Test: DexArchiveBuilderDelegateTest
Change-Id: Ief036ca0ffbdc171491f472e532b385d8e50e7e4