Desugar library code when library is minified
If language desugaring is enabled for library project
make sure to desugar library code when building DEX.

Previously, if minification was enabled AGP reports
R8 as desugaring tool and D8 will not be used for
desugaring. This is an issue because R8 does not run
when building DEX for production library code.

Bug: 154218507
Test: DesugarLibraryAndroidTestTest
Change-Id: I1ea7ef2a5de82d4d56fe8c4f8f802a73145402ae