New DSL: Remove CommonExtension return type parameterization
While this change looks messy, it actually should have no effect on
binary or source compatibility.

This is done to prepare the codebase for removing the parameterization
entirely, which means removing it from the "block" configuration
methods, which is a breaking change for Kotlin build script users that
specify types explicitly.

The new Internal* interfaces are introduced are simply to allow the
types for the kotlin delegation to work correctly without having to
duplicate manual delegation for each of the four implementations.

Bug: 140406102
Test: Existing DSL tests.
Change-Id: I5ee3c983a404ffd934edcfd15771af571caa182c