SDK build service and global sync issue service
Introduce SdkComponentsBuildService and GlobalSyncIssueService.

The first one is used to load all SDK components,
and to make them shareable across build. Service
will be instantiate when Provider holding its value
is queried. Suppliers have been replaced with lazy
properties and Gradle Providers, which required some
changed in the SDK-loading classes.

The second build service is build-scoped sync issue
reporter, which is currently used to report issues
when loading SDK components. This is necessary in
order to match the sdk build service lifecycle. Issues
collected are queried for from the model builder, but
they will be reported only once. This is to avoid any
duplication.

Services are not used in this CL, next CL will wire them
into the build.

Bug: 152321793
Test: SyncIssueReporterImplTest + existing
Change-Id: Ia0d68ce94e84d0789d30ae2c9dcfcf3f26c003e2