Kotlinize command-line parsing classes
Bug: 155099856
Test: Existing tests pass

This is in relation to fixing the referenced bug. I'll need to update
the tokenizers to break out the zero allocation parsing logic from
the allocating token-forming logic so that I can replace the
token-forming logic with something that performs less allocation for
the scenario in the referenced bug.

The logic for separating those will be better represented as Kotlin
closures than Java classes.

Change-Id: I6503584d80ee3c3188093c8cbde3c8577820842b