152539667: AGP 4.0.0-beta03 Lint Error [MissingClass] false positive
This CL fixes the AGP integration with lint which computes
the classpath to set up the environment with. The code
which iterated through the builder model to find the output
folders, would check to see if the output folder exists,
and if not it had fallback handling used in libraries where
there is not a variant match.

Anyway, this "does output exist" check was looking at the
primary output folder, which is the java compilation output.
In the test project associated with this bug, there were
no Java sources, only Kotlin sources, which means that the
javac/debug folder does not exist. Therefore, lint did not
proceed to add the *other* output folders, notably the
kotlinc output folder, which results in this bug.

We have had Kotlin-only projects before, but AGP has always
created the javac output folder to write its own BuildConfig
class into, so this check would succeed. However, in this
test project it also turns off the BuildConfig generation,
and that's what triggers this scenario.

The fix is similar to what has already been done in 4.1
after the lint model conversion: we treat all the class
outputs uniformly, not just the primary folder.

Fixes: 152539667
Test: Integration test included
Change-Id: If6d00ce1adfacd58898b64d4fd7f7256228ab390