Refactor command-line parser to separate allocating side from ...non-allocating side.
This logic is a hot-spot for parsing large compile_commands.json
files produced by CMake (see b/153964094#comment7 for a call-
stack).

This CL factors the parser into two parts:

1) The first part is the parsing logic itself. After this CL,
the parsing part has no per-character allocation.

2) The second part constructs a List<String> of tokens that is
expected from the code that uses the current tokenizer.

The Windows side of the parser needed to be reworked because it
had logic that walked backwards and that's not amenable to
forward-only streaming. I think the result is a simplification
and is easier to understand. There are extensive unittests that
cover the behavior of this parser.

Actually fixing the compile_commands.json performance bug will
happen in a future CL.

Bug: 155099856
Bug: 153964094
Test: Existing unittests
Change-Id: I2ab353960a62b2e9644c154b6fdfb46d1eeecf91