Duplicate coverage master ignore logic into standalone filter
This lets us use the master ignore list in other places (specifically
for baseline coverage). This is just a temporary duplication. The logic
in the XML->LCOV conveter will be removed in the course of the baseline
coverage integration. I'm just splitting it out like this to make
following the process across the commit chain easier.

Test: N/A
Bug: 153913236
Change-Id: Ic0f9f85c1aefa992f60b3a2e07a3dddf10a227af