More v2 model clean up
Remove extra appBundleBuildOutput and move this into
BundleInfo, though it may be redundant with what we
have there already.

Removed DataBindingOptions which is only used by the DSL(!)

Introduce TestInfo that contains the old TestOptions plus
some things from AndroidArtifact (goal is to reduce passing
all this if this is not a test variant)

Minor clean-ups in project, base artifact, especially kdoc
to make them more kotlin friendly.

Bug: N/A
Test: N/A
Change-Id: I495e43f2620d4a1697b27330405df915122c340b