DSL: Expose ABI filters as val Mutable
This reverts commit a36f7c0696f82a917ce8b8be0c1692fb89769c22.
(Reapplication of f880dc97c4b8190eca6d55b1554f81b34126f42e)

It was previously nullable that would make this problematic,
but consumers intended to treat empty collection the same as
null, so now it is not nullable. This change fixes various
inconsistencies where some places did accidentally.

Not a source-level breaking change for kotlin script users
as the existing setter before 4.1 was a builder-style one
so did not make the field mutable from a kotlin perspective.

Reason for revert: Reapply with fix on Studio side.

Bug: 140406102
Bug: 156169950
Bug: 157151941
Test: Updated API expectation files
Test: Existing AGP integration tests, e.g. NoSplitNdkVariantsTest
Change-Id: I6de394be8335a672243504190068cb064a46629d