Move exec and javaexec into Cxx service
This is to separate the concern of starting process out of the
ExternalNativeJsonGenerator interface. The reason is that
future generators may not need to execute processes to produce
a model.

Separately, I think we need to inject the exec service into
ExternalNativeJsonTask in the future to unblock gradle instant
execution.

Bug: 153964094
Test: Should be no behavior change
Change-Id: I6533e5a76059ef2f226296eef19ec89d7713dd3a