Prepare DexMergingTask for incremental support
Currently the inputs to the dex merger tool in DexMergingTask are dex
roots (directories or jars containing dex files).

To support incremental dex merging, we will split the input dex files
into multiple buckets, and invoke the dex merger tool (D8) once for each
bucket in a worker.

This CL refactors the pipeline so that the inputs will now be a list of
dex files (or dex jar entries) of a bucket. The actual bucketing and
incremental support will follow later.

Bug: 132615300
Test: DexArchiveMergerTest, DexMergingTaskTest, and integration tests

Change-Id: I4668b66ab61bf21839da4ca93d733efff641ad07