Avoid using buildFinished listener for analytics
This CL introduces ProfileCleanupBuildService that is
used to finish the analytics profile output. AGP is not
relying on buildFinished which is incompatible with
configuration caching. Having fully analytics that is
fully compatible with configuration caching is tracked
in http://b/157470515.

Also, gradle.projectEvaluated is used to capture
relevant info when project is evaluated (previously
build listener was used).

Using task listener on the task graph uncovered a
concurrency issue when recording memory samples, so
this CL fixes that as well.

Bug: 154742527
Test: BasicInstantExecution
Change-Id: I92a504340b951fb09960dcc32695212f9bf70ccf
(cherry picked from commit 096c1ec035250d47052052e7239d62220b5cc9e6)