DSL: Improve packagingOptions setter semantic.
The previous copying-getter semantic lead to uses of the form

        Set<String> myExcludes = getExcludes();
        myExcludes.add("example");
        setExcludes(myExcludes);

However, that would result in the setter clearing the collection,
as the set passed to the setter is the same instance as the main
collection.

Now make a defensive copy first.

Bug: 140406102
Bug: 156169950
Test: Extended PackagingOptionsTest
Change-Id: Id79eb62216bc309aaeae57a4b54576a46659b589
(cherry picked from commit 9ed233d78dcab2f938ce8d45b3584ad4cc4b9574)