Fallback to IPv6 if adb server not found in IPv4
It's possible for the adb server to start on the IPv6 loopback address
::1 instead of the IPv4 loopback address 127.0.0.1. This CL changes
AdbSocketUtils#openAdbConnection to try both IPv4 and IPv6 loopback
address when connecting to the adb server.

AndroidDebugBridge#getSocketAddress has been patched return a known
working SocketAddress. If the attempts fail it falls back to the old
implementation.  The methid has also been deprecated because it
returns either IPv4 or IPv6 loopback address but not both. Consumers are
encouraged to use AndroidDebugBridge#openConnection when establishing a
socket to the adb server.

AdbSocketUtils#openAdbConnection has been moved to
AndroidDebugBridge#openConnection because AdbSocketUtils is in an
internal package while opening a connection to ADB is generally useful.

AndroidDebugBridge#openConnection will not try both IPv4 and IPv6
loopback address when connecting to ADB.

Bug: 157173130
Test: No existing tests broken.
Change-Id: I690dab1afa84d0cbaa594ec750136cf06ef0eccc
(cherry picked from commit 0c02208602ff7b256ff93970549f075a4329968b)