make Single and Multiple subtypes of Artifact.
having them separated from Artifact hierarchy made it difficult to store them
as variables and hold both types. Therefore lots of internal code was
using the internal type that combined both Artifact and Single into a single
type.

With this change, it will be possible to publish external public types for
inter-project consumption which will remove the need for republishing.

Test: Existing
Bug: 158228529
Fixes: 158228529
Change-Id: Id37157f7713c5104cc867d2549b80514621b0cda