More implementation of ModelBuilder v2.
Implements:
- the AndroidProject object minus the variants
- the issue reporter.

this also register the new model builder in the
model builder registry.

Basic Unit tests for the issue reporter model
and one integration test for AndroidProject.

Have to duplicate a bit of the infra to
query model, as well as the modelcontainer result
in order to handle v2 models.
Also implemented a model dumper to be able to more
easily validate the content of the model.

Bug: N/A
Test: updated ModelBuilderTest, new BasicModelV2Test
Change-Id: I75dc9baa496522ab114da59be93fe872191dfacd