change output.json file name to not collide
we supported moving up in AGP versions with regards to the
output.json handling but we did not anticipate that users keep
moving up and down between branches and the file handling was
not done well in 3.6 and below for a file generated in 4.0.

The simple thing to do is to rename the output.json so they
can live side by side and be ignored by respective plugins.

Test: existing
Bug: 156556231
Change-Id: If16d686ba1540e18a66eb70af000415b1bbe50cf
(cherry picked from commit 5ea305d9eb6d3cf398f86307d41e7850e144558d)