Change lint model serialization adapter to use folders
A lot of the lint serialization methods used the
path to the module.xml. This made sense when the entire
model was serialized to a single file, but now that
it's not, instead of using the parent of the module
file, change the API to take a folder as the source or
destination instead.

Test: Existing
Bug: None
Change-Id: I26eedb2fc2ba40d5511599c78f8850b27d9c6a65