Reduce the number of AndroidX substitution rules
Instead of applying the rules to resolvable configurations, we now apply
the rules just before the configurations are actually resolved.

This greatly reduces the number of rules, as not all resolvable
configurations are resolved in a build.

Example effect: In a small project with 1 app module, 1 Android library
module, 1 Java library module, we have reduced the number of rules from
209 to 18 when running `assembleDebug`. The configurations where the
rules are applied include:

app:
  :app:debugCompileClasspath
  :app:debugRuntimeClasspath
  :app:debugAnnotationProcessorClasspath
  :app:debugReverseMetadataValues
  :app:kotlinCompilerClasspath
  :app:kotlinCompilerPluginClasspath

Android lib:
  :lib:debugCompileClasspath
  :lib:debugRuntimeClasspath
  :lib:debugAnnotationProcessorClasspath
  :lib:androidApis
  :lib:kotlinCompilerClasspath
  :lib:kotlinCompilerPluginClasspath
  :lib:lintClassPath
  :lib:lintPublish

Java lib:
  :javalib:compileClasspath
  :javalib:annotationProcessor
  :javalib:kotlinCompilerClasspath
  :javalib:kotlinCompilerPluginClasspath

Bug: 158846149
Test: Existing JetifierTest
Change-Id: Ia697fda1d9d63c663e3d24a1bd349030df511cc5