Add caching to startup instrumentation
When the IWI agent attaches on app startup, it instruments specific
framework classes to allow overlays to be loaded.

The most expensive portion of this operation is using dexter/slicer
to rewrite the target classes. This can incur several hundred
milliseconds of additional app startup time, potentially negating the
benefits of a faster deployment.

Because the target classes are framework classes, they will not change
between app runs. Thus, we can limit our impact on app startup times by
caching the result of the instrumentation on the first run and only
performing a dexter/slicer call if no cached class is available.

Also updates FakeDevice to support testing startup agents, in order to
allow the transform caching to be tested.

Bug: None
Test: Manual + DeployerRunner.testAgentTransformCache
Change-Id: I1b932f3703dc36353fb77740cab37642a2886f37