Update PackagingOptions DSL
This change adds new blocks to the PackagingOptions DSL for
dex, jniLibs, and (java) resources.

In subsequent CLs, this new DSL will be wired to new VariantProperties
objects and then to tasks.

Also in subsequent CLs, assets and (android) res blocks will be added
to the PackagingOptions DSL with the ability to disable compression for
particular assets, res, or resources.

Bug: 129428538
Bug: 149770867
Test: existing
Change-Id: I16f6b955e9809705eca7099771dab62b1d1cbacf