Update lint formatting
A while back, we switched to lint being formatted with
google-java-format (for Java files) and ktlint (for Kotlin files).

However, since then, there have been a few changes to those tools
which affects the formatting, so the files are no longer matching
the style created by those tools. The whole point of reformatting
the codebase was to have an automated and simple way to canonically
format all changes, and currently, after applying those tools I
need to selectively revert diffs unrelated to my change. Therefore,
this CL updates the lint codebase to again reflect a full format
operation applied by those tools.

For Kotlin, the changes for ktlint are:
- Changes to hanging indents
- Breaking up longer lines more aggressively, particularly
  in method declarations
- Forcing multi-line raw comments to open on their own line
- More aggressively ensuring that a closing ")" in a multi-line
  argument list is moved to the next line

For Java, the changes for google-java-format are:
- Enforcing a space after "//"
- More aggressively breaking up longer lines in comments
- Changing hanging indents in some cases

In some cases (particularly the Java formatter's line breaking in
comments) the changes aren't really an improvement, but there's
an advantage to having this formatting be fully automated.

Versions:
google-java-format: 1.7
ktlint: 0.37.2

Test: Compiler + existing tests
Bug: None
Change-Id: I433a5420e4d804cb5d11642427d90b3d5ce37a95