Disable mlModelBinding flag for test variants
Regressed from ag/12204287 which changed the default value for
overriding from 'false' to 'null', so now we need to specify that
explicitly.

Also re-enable MlModelBindingConnectedTest to prevent such
regresssion in future.

Bug: 162323560
Test: broken test from ag/12204287 is fixed and manually verfied test
variants (both android and unit test) in a demo project can be built
without issue with this flag enabled in build.gradle.

Change-Id: I793ee0538ca206d13e349e5e5a07a0c9436c31f0