Copy outputs from correct directory in studio_mac
When using bazel info to retrieve the bazel-bin directory to copy the outputs,
we need to pass the same config flags used to test/build otherwise we are
copying the wrong binaries.

By not passing the config flags, we were fetching the binaries from fastbuild
(the default compilation mode) instead of the optimized ones (-c opt).

Test: manual/local
Bug: n/a
Change-Id: I81d17e360b4890c44560ea273611e8b7fe4e6d4e