Fix support for dialogs
After I89f1d780, trying to render windows with origin not at the canvas
origin caused the skia parser to crash.
This fixes that by including the offset of each window from the canvas
origin, and (in this commit) using that to correctly set the canvas size
in the skia parser. This involved creating the canvas before playback
starts, rather than lazily as before (which is nicer anyway).

In the companion change in tools/adt/idea, the bounds of requested views
are transformed from screen coordinates to canvas coordinates.

Test: Tests for the skia parser will be added in a separate commit.
Fixes: 167705169
Change-Id: I66c33e637b6e76794cd4c2b07f841cb1ab75a6ca