168261940: Prevent configuration cycles
This is a hotfix for a scenario in 168261940 where lint can
create a cycle in the configuration hierarchy (specifically,
if you use a --config named exactly "lint.xml" which sits
in a project directory; that isn't the case in b/168261940
but I encountered it when trying to reproduce it, which makes
it hang looking up configurations. This CL just breaks the
cycle, so it's possible that you end up with the wrong
inherited values from lint.xml, but that's preferable from
lint hanging. An upcoming (much bigger) CL will fix this
properly.

Test: ProjectInitializerTest#testManualProject (updated)
Bug: 168261940
Change-Id: Ibfbe26f41363da8431dc41b99a06bfb2a4ba71ba