Update OverlayId to track removal of overlay files
This ensures that the files tracked by the OverlayId are always in sync
with what is actually in the overlay on the device. Currently, if files
are deleted from the overlay, every subsequent OverlayId object will
incorrectly indicate that those files are still present.

Fixes: 169262624
Bug: 169250554
Test: OverlayIdTest.java
Change-Id: I2ed9f12de0b899a5c65789ad6e06d64b468fa1f0
(cherry picked from commit c2e7ac3c1983afd0981335812ed92f89232ee909)