Clean NewIncrementalTask outputs when not incremental
This change adds an explicit output cleanup to NewIncrementalTask
when inputChanges.isIncremental() is false.

Gradle should be doing this cleanup automatically, but there are cases
when it doesn't. This seems like a Gradle bug. This change is a
workaround until the Gradle bug can be identified and fixed.

Bug: 169701279
Test: androidx-studio-integration
Change-Id: Idc8377c4ff26627695beaa086db916f9ec42f376
(cherry picked from commit bc19a2f084dcbdc2f88eed60ff6d81857b295434)