Stop using reflection for IWI resources
Previously, the SDK for API 30 was not available to build against, so we
had to use reflection to access the classes needed to make IWI resources
work. Since it is now available, we can just use the classes directly.

Using reflection meant we needed to manually find/call methods using
getMethod and getDeclaredMethod. We made an assumption that a particular
parameter would always be a Resources object, but in the related bug, a
derived class - CompatResources - was passed. Because of this, our call
to getDeclaredMethod to resolve the method failed (we would have
needed to call getMethod instead).

Now that we do not use reflection, the VM handles method resolution for
us, so we can just rely on polymorphism again.


Fixes: 170709330
Test: Manual
Change-Id: If176e38a44d1ced21e9e64950fd5a085a7663e29
(cherry picked from commit 2485acb3882bc02756ceb2bedee8a457ce96f245)