Switch lint unit tests to using a client factory
Lint's testing support had a way to pass in a custom lint client.

However, lint sometimes needs to create additional clients, for
example to run the checks twice, once with UInjectionHost turned
on and one with it turned off -- and in an upcoming CL even more
scenarios. Reusing the same client runs the risk of not testing
the real behavior, since keeping state around might help or hurt
the second use of the client.

To fix this, instead of registering a specific client instance,
they now implement a callback to create a new client, which lint
can call repeatedly.

Test: Existing
Bug: None
Change-Id: I327dd21c9ef4a2afad8fa00fc924f6498f8524aa