Update lint fix data map to be persistable
Most lint quickfixes are described as document edits or XML operations
as well as logical operations like reformat or add import.

However, there are some cases where the fix is more complicated and
lives in the IDE codebase, and the lint check simply passes a map of
keys+values to the IDE for interpretation. In some cases, this would
include data like PsiElements.

This CL updates the quickfix API and usages to only use	persistable
data (along with updates on the IDE side.)

(Note that the changes to LintFix's data map is incompatible, but that
should be fine since no external lint check would have any use for
this special fix type, which is really a protocol between lint checks
and special handlers on the IDE side.)

Test: Existing (and on the IDE side)
Bug: None
Change-Id: I2e8b1a9c26dba2b94f72aeb346003a8758721c2d