Improve ExtractNativeDebugMetadataTaskTest
Use an unstripped .so file to test extracting native debug metadata.
Previously the test was using a stripped .so file, so there was no way
to verify that the native debug metadata was being done. Now that it's
using an unstripped .so file, we're able to at least test for the
expected relative sizes of the bundle entries.

Bug: 169266007
Test: ExtractNativeDebugMetadataTaslTest
Change-Id: I75573ba34d913638909cd794d2aab0c885ce1aaf
(cherry picked from commit 9cfdfef41da331684a5c75efd8bbfd1472eae43d)