Fix analytics service not configured in composited build sync
This CL changes the way of checking if it is a model query or a build
with task requests. Previously, we check task requests on
startParameter. But in some cases, the startParameter has a single task
request with empty args and we should consider this case as a model
query.

StartParameters.taskNames aggregates all args from task requests which
is a better option for us to check if it is a model query or not.

Test: manually
Bug: 175159902
Change-Id: I6af10176ae92b5c87a481073b373e8a4cc97d71f