Lint: hotfix for b/159733104
Whereas Change Ibff843abc adds an "opt-in" workaround for b/159733104,
this change adds a workaround for everyone automatically. Thus it
is slightly riskier.

This change resets the FacadeCache in the Kotlin compiler before each
analysis. This is needed because Lint uses the same UastEnvironment
for multiple analyses, and stale entries in the FacadeCache can
lead to NPEs.

(cherry picked from commit 9dfd18dcccb5c8692d314e1378bfcb361cd99eb1)

Bug: 159733104 (23 upvotes)
Test: tested manually on the Tivi project
Change-Id: I3c7422fa43dfa7619f3f864d66ef9b9e289bdc02