Add message chunking support to skia parser
This allows larger SKPs to be sent to the parser.
Also split up the response, with images being sent separately. This
doesn't get the response size below the default limit in the normal
case, but is helpful in getting there eventually if we want to.

Also removed the "image too large" functionality in the agent since it's
not necessary anymore.

Test: Added a test, but mainly adding an integration test in adt/idea
Bug: 146165415
Change-Id: Ia81502afccc88da1b4b83e0e0eb9ad991c7bf8d0