Add aspect to calculate disk size of data dependencies
This is useful to know how much data a test uses at runtime, and breaks
down the size per data dependency.

The aspect uses no-remote-exec, because setting up a local linux sandbox
with symlinks is near instant but transferring gigabytes of data
dependencies for remote execution is not.

The datasize_aspect config is added to studio_linux.sh
to enable it on CI builds. Eventually these .datasize
artifacts can be collected in post build processing to provide
size comparisions between builds.

Test: NA
Bug: 175231442
Change-Id: I9523870ef000d4cc82359d076150e0adab266914