introduce versioned NDK from SdkComponents BuildService.
Similar to sdk, the ndk can be versioned per module but the
BuildService was using a singleton so first module to initialize
it wins.

Added versioning input nested object for convenience inclusion
in Tasks requiring a versioned ndk, and removed old APIs and old
singleton property from the build service.

Test: unit test and integration test for regression.
Bug: 169931635
Change-Id: I578277c6469a762bfceebcaafd1655aaddd77ee2