Add a new pre-AndroidProject model for versions
The idea is to provide model versions so that studio
can figure out whether it can open the project, and provide
guidance if it cannot.

For now versions are 1.0 for all models.

Updated the fixtures to handle this. Now the model container
does not contain a single model but all possible android models
depending on the query. This allows syncing both AndroidProject
and VariantDependencies in a single query for instance which
will be faster.

Bug: N/A
Test: new snapshot for the new model + updated tests for new API.
Change-Id: Icc6c6bd12d8c6613040068c46b3599ccd9fb6dec