Fix a bug in TPD symbol exporting.
The symbol file was using fput which terminates at the
first \0 encoutnered in the string. This resulted in
truncated symbols. A second issue was encountered where
on windows the files were ASCII instead of Binary
as such the "B" is required when opening the file to hint
the output mode.

Test: manual, attempted an integration test it get really
cumbersom with 3 platforms and a dependency on the NDK tools.
Bug: N/A

Change-Id: I7fae5dd68622483a9518499d498828f55fe4778d