Add path variables to lint persistence
The lint model already had support for path variables, to make the lint
model XML files relocatable (which is necessary for caching). This CL
generalizes this support and hooks it up to the partial analysis
persistence files as well. It also adds a command line flag for
providing the path variables (such as the Gradle project root, and the
~/.gradle folder, and so on) to lint.

Finally, it adds this new flag and a bunch of other "internal" ones used
by the Gradle integration to an advanced options category.

With this, AGP can provide for example

--path-variables GRADLE_USER_HOME=/Users/tnorbye/.gradle;PROJECT=/Users/t...

and lint will write its persistence files with the variable names as
prefixes instead of the full paths. When the same path variable names
are provided for reporting (possibly with different actual path
substitutions), the files will be deserialized in the reverse way, which
should allow internal results from analysis to be cached.

Test: New and existing unit tests
Change-Id: If8fc016e39bb41734b6e10706c76e1d3da1ac77d