Revert "Stop running targets tagged as `very_flaky` in coverage"
This reverts commit 425cedb8ebcd5c5bc2fa04db0b9411108bf11fdc.

Reason for revert: This change is a no-op, which is good because if it had its intended effect it then coverage post-processing would have exploded. Since the added build tag has no semantic effect it should be removed to avoid confusion by later readers.

The coverage bot invokes a test_suite and --test_tag_filters will only check if the test_suite has the appropriate tags rather than whether tests in the test_suite do.

In the future, please do not change the coverage bot's behavior without my +2 unless it's an emergency and I'm unavailable.

Change-Id: I61206b85f2647db9765af67a583402d7be70328d
Bug: 153972155