Single variant publishing for library and application projects
With this change, we add the new publishing dsl for users to publish a
single variant to a component. For library projects, the variant is
publish as an AAR artifact. For application projects, the variant is
published as an APK or AAB. Note publishing the variant as APK and AAB
to the same component is not supported. By default, we publish the
application variant as AAB.

With the new publishing mechanism, we are not going to create software
component automatically and the component name for APK/AAB is same as
the variant name instead of something like variantName_aab or
variantName_apk.

In order to offer a better upgrade experience for users, we still create
"existing" components if users don't opt in this feature(they don't use
the new dsl and don't use the new introduced flag to disable automatic
component creation). When developers don't opt in this feature, we raise
warnings about the behavior change so that they can migrate to use new
publishing dsl before the support for old behavior is dropped
completely.

Bug: 150132264
Test: existing + LibraryPublishingTest + AppPublishingTest
Change-Id: Iabea55a7692c8736f13c6337c61e6a79401489da