Mark suppressUnsupportedCompileSdk as stable
The android.unsupportedCompileSdk property is suggested in the warning
by the Android Gradle Plugin as a way to suppress that warning.

Marking it as stable means that build authors don't just trade one
unsupported warning for another, especially as
suppressUnsupportedOptionWarnings is not discoverable at the moment.

For now android.suppressUnsupportedOptionWarnings is left as
experimental as it can be used to suppress its own warning along with
the other experimental options used in the build, as follows:

android.suppressUnsupportedOptionWarnings=android.suppressUnsupportedOptionWarnings,[...]

Bug: N/A
Test: None
Change-Id: Ic00235a97e7e1d14cd97e242b6bc60e747a90edc
(cherry picked from commit acd11fc1ec8d5f9f0b625d4121dcd01cd01fbfed)