Add workaround for Windows classpath jar
When the length of a java_binary's classpath is longer than 7000
characters, the launcher on Windows generates a classpath jar. This
classpath jar is incorrectly generated when run remotely as part of
ctx.actions.run.

This change uses the --singlejar flag to force Bazel to
build a standalone jar and run that instead so that the classpath jar is
not used regardless of the classpath length.

Bug: 182291459
Test: Existing tests pass
Change-Id: I2b8edb467b8d2615134a6c21751a3d3236b3b056