[lint] Defend against unparsed Gradle dependency versions
The version parser for Gradle dependency specifications is not complete,
and also does not report parse errors for versions it does not
understand.  In practice this can lead to misunderstandings where the
version of a specification is parsed as "0", leading to false-positive
lint errors.  Defend against the zero Gradle version in version
comparisons in Gradle detector.

Bug: 187955701
Test: existing test enhanced
Change-Id: Iaacf8cdff1a2d8fa4a9f266152addb0007f82ecf
(cherry picked from commit 9f34add772620e5cf1c26aeae703438eeeef2124)