Lint: ensure system properties stay set
The Gradle daemon resets system properties between builds; see
org.gradle.launcher.daemon.server.exec.EstablishBuildEnvironment.
So, when Lint run in-process in the Gradle daemon, it is
insufficient to set system properties from a
static initializer that only runs once.

This change also removes a couple system properties that
are redundant because they are now set by the Kotlin compiler in
setupIdeaStandaloneExecution().

Bug: not filed; this is a follow-up for Change I9e8309fb1
Test: existing Lint integration tests
Change-Id: Iebdad6c863cccb7098aba5a85f36b61f90381c39