Warn about using compileSdk with a string in build.gradle files
We now create build.gradle files with "compileSdk <number>" to specify
an SDK. If you just change the number to "android-S" to use a preview,
which is how you'd do this in the past, then Gradle will emit a
confusing error message. This change makes the Gradle check flag
combinations of "compileSdk" and a string when you're editing in the
editor and suggests switching to "compileSdkPreview". It also updates
lint to apply its previous checks for compileSdkVersion to compileSdk as
well.

Test: Unit tests included
Bug: None
Change-Id: Iddf6127362840f820d3b68bc8f64adf0979ba8d3
(cherry picked from commit 3a87fc01d651c56f474d92ab0bf77f10f5c007ce)