add variant API to set signing configuration.
The Variant API SigningConfig had properties to set the various
enable flags related to V1/V4 signing but nothing related to the
keystore to use.

Although it is not offered to reset DSL information directly, a
convenience API is provided to identify the name of the DSL
SigningConfig to be used to sign the variant's artifacts.

Test: added recipe
Bug: N/A
Change-Id: I49bb71513d6988026f85cb3827ebb6aa0c33965f