Remove ParentLibsTest & ParentLibsConnectedTest
These tests check that the build succeeds when a subproject is located
outside of the root project directory.

However, this support is deprecated in Gradle 7.1 and will be removed in
Gradle 8.0 (see
https://docs.gradle.org/7.1-rc-2/userguide/upgrading_version_7.html#deprecated_flat_project_structure).

This commit removes these tests as they are no longer applicable. (We
could also remove them in AGP 8.0, but it's probably safe to remove them
now to save resources & reduce work when upgrading to AGP 8.0.)

Bug: 189931893
Test: Removing them
Change-Id: I7cdad3d63363597abe14df2140b1b3766827cffb