Get rid of the lint progress printer
This was added back when lint was typically run as a separate lint
command, pointing to directories containing many modules which it was
iterating on itself, or even folders containing a lot of projects (such
as all the Android samples), and displaying progress was useful since
there was no surrounding build system to do it. Now, with Gradle, this
is typically just noise.

Bug: None
Test: Existing tests updated
Change-Id: Ie2a0369815d23170c255bacf49a4f387df80ff09
(cherry picked from commit 8cba60c327f1ac53313eb5b1a95b1b1c75c8095a)