Flag absolute paths in lint error messages
Including an absolute path to a local file in a lint error message is a
Bad Idea™ -- it will not play nice with remote caching, or baselines,
etc. In lint unit tests catch cases where a lint check does this by
searching for paths in reported error messages, checking if they're
existing files on disk and if so, fail the test. The regression test for
this also revealed a few problems with the existing expected-exception
handling in the test infrastructure which this fixes.

Bug: None
Test: Included unit test
Change-Id: Iaee4377e03de02ed5d8d5fc46c24de86f23741c4