Fix androidTestNamespace in the model.
The model namespace values were fed via
variant.namespace which is fine for apps and libraries
but not AndroidTest.

Due to legacy reasons AndroidTest uses the appId of the
variant to generate the R class. This is wrong and will
be fixed later but this is a source level breaking change
that requires upgrade support. (b/176931684)

Because BuildConfig for AndroidTest use the proper namespace
value, we cannot just change variant.namespace to return
the correct value (based on appId) and then feed that to
the model.

For now (AGP 7.0 and possibly all 7.x), this commit does the
following
- Keep variant.namespace to return the value based on the manifest
  or the DSL namespace attribute only.
- Keep BuildConfig to use that value to generate its class
- Keep the R class to be based on the appId in some cases
- Make the model return the namespace of the R classes rather than
  the value returned by variant.namespace.

The goal is to ensure that Studio can use the namespace value
from the model to create the light R class. BuildConfig is not
create via light classes in Studio (the IDE load the generated
classes by the build), so this is fine.

Fixes: 191770358
Test: added new model v2 tests for namespace customization
Change-Id: Ie118a8df49b5e892dd1f521588a20634d6320ee0