187341964: Android Lint VectorDrawableCompat check warns incorrectly
This CL changes the way we analyze this. Instead of looking up icons
lazily via the resource repository (which involved actually reading XML
files later), we record whether drawables are vectors when we encounter
them, and whether rasterizing is turned off in the current project.

Also updates the model mocker to include the project name in the model
name such that it's easier to tell from a unit test what the model is
(and in this test we want to make sure that we're picking the right
module's build.gradle file to point the user to.)

Fixes: 187341964
Test: Existing and new unit tests
Change-Id: Icd630b9cf015893e0fe15d48f4c28641ea65c9a2