192870410: Improvements to lint's bytecode verification of lint jars
Lint now performs some bytecode verification of lint jars before
loading them, to both make sure we properly flag old and incompatible
lint jars from AAR dependencies but at the same time don't create a lot
of noise about perfectly fine older lint jars that are only using the
stable API surface.

However, if a detector is *itself* packaged in one of the namespaces
lint considers part of its API surface (such as com.android.lint) then
this mechanism breaks -- and that recently happened to various lint
checks written for the Android platform. This isn't specific to
com.android; lint also considers PSI and UAST as part of its API
surface and if a lint check happeend to be in one of those packages the
same thing would happen.

This CL improves this situation in a couple of ways:

* `com.android.internal` is now a specifically reserved package prefix
  for lint checks, so the platform lint checks can be moved there
  without having to be placed in misleading packages.

* Lint now specifically validates lint jars for this packaging issue
  such that it doesn't show up with a confusing error message. This
  also helps catch cases where you've accidentally packaged API jars
  into the lint jar, such as the kotlin runtime library.

This CL also fixes a few related things encountered while fixing the
above:

* Bumped the Lint API level for 7.1 and updated some tests to handle
  this more gracefully for next time.

* Found a bug in the binary test file generator code which computes
  checksums (it was including the checksum for all compiled files
  instead of just the ones for the current test file)

* Added checksums in all the missing places in various unit tests such
  that the deprecated checksum-less method is no longer used.

Bug: 192870410
Test: Existing unit tests plus various new unit tests included
Change-Id: I513f28bf6ff14655dc0e4a13805d59b01d4a086a