Moves gmaven tests back to .zips
This is a fast test that validates the structure of the generated gmaven,
and is the only one that uses gmaven built in bazel.
This move is in preparation of switching gmaven to be built with gradle+bazel
and the output of that is a zip.

This reverts part of commit c84a9dc6dc1608c9ad2088f9c73eba0d0f73fa80.

Test: this is it
Bug: 190268212
Change-Id: I32cecb7cccb461643165a0e2359a6890fc168e9c