Change native library instrumentation target
The current instrumentation point for native libraries rewrites the
search path in the DexPathList for the application classloader, which
allows calls to System.loadLibrary() to prefer the overlay native
libraries over the installed libraries.

However, the linker does *not* use this path to resolve library
dependencies; instead, it searches inside a linker namespace created
when the application classloader itself is created. Thus, our current
instrumentation causes the linker to be unable to find library
dependencies inside the overlay.

To fix this, we move the instrumentation point to right before the
ClassLoader is created, modifying the native library search path
parameter passed to ApplicationLoaders.createClassLoader(). This causes
the linker namespace *and* the application classloader to be created
with the correct path.

Fixes: 186806270
Test: Manual
Change-Id: Ic3e3404ef50eb1b71363519f9bcd6b975a56efca