Remove --keep_going from CI scripts
If bazel encounters an error, prefer to fail immediately. Failing tests
do not count as an error, so we still see all test failures. Build
failures are infrequent, but when they happen we should provide feedback
as soon as possible. Using --keep_going delays CI results, because
unrelated and long-running tests continue to run.

Bug: NA
Test: NA
Change-Id: I0a48506d7a5208c99703dd86549d202a188ece8c