Fetch on any non-null/non-empty MAVEN_FETCH value
Bazel also uses the value of MAVEN_FETCH environment variable when it
caches repository rules. Therefore, we have a problem when:

1. User fetches some artifacts using:
    MAVEN_FETCH=1 bazel build @maven_tests//...
2. User deletes the fetched files manually, e.g.,
    cd prebuilts/tools/common/m2
    git clean -xfd .
3. User tries to re-fetch using same command in (1).

But, Bazel won't re-fetch, because the repo rule hits in the cache.

Adding entire m2/repository/** as an input to the repo rule can ensure
any added/deleted file invalidates the cache, but that can cause
over-invalidation (e.g., increased cache misses), so I discarded
that option.

An alternative is to use:
   MAVEN_FETCH=$RANDOM bazel build @maven_tests//...
Which we can put behind a script.

Bug: n/a
Test: n/a
Change-Id: I82235175fcfc0f3612295adef38becaf459f0af0