Write AGDE version to App Metadata File (from StringOption)
Android Game Development Extension is built on top of
Visual Studio, and orchestrates builds/testing/etc. for
developers making use of C++.

This change adds a StringOption which allows AGDE to
inject a version# at build time, which will be included
in the final Bundle/APK, to facilitate metric collection
on the Play store.

Bug: 190107180
Test: AppMetadataTaskTest (in gradle-core)
Test: AppMetadataTaskTest (in integration-test/application)
Change-Id: Ib3d2812c17252fe1c48b138b965f7753fe372c36