[UTP AGP] update emulator close to use adb.
Use adb as the primary action to close the emulator. When the emulator
process is killed directly, it does not necessarily allow for snapshot
creation. In order to better facilitate snapshot creation (and to allow
more consistent closing of the emulator from the setup task.

This is done in the same way we presently delete devices in the
GradleManagedAndroidDeviceLauncher in UTP. i.e.
1. look up all serials through adb
2. search for a unique id for the device (which was passed into the
emulator) through adb
3. run adb kill on the given serial
4. destroy the process as a backup if the serial could not be found.

Also added gating to the avd detection within AVD components, to
potentially help with a duplication issue with AVD Manager

Test: updated testing.
Bug: 199375443 199374877
Change-Id: I0f281b7c222f49df8e260fa3bfe2f906af847d68
(cherry picked from commit d2f05da630b582a40470f66dff6cf5a08c322ddb)