Fix the issue second Google Pay Activity is not created
This happened because the Activity was created in the sub folder
($packageName.activity), thus uniqueness of the Activity was not
correctly checked.

This CL changes the package of the Activity to the same package as the
application's root package and make the uniquness check of the Activity
work correctly.

Bug: 199817138
Test: Manually followed the steps in the bug for both Kotlin/Java Activities
Change-Id: Iecde5bb92ddb0b05aacc8a550f1ba815a144fa1d
(cherry picked from commit 4f8d101ee388a78f2bd666faf067f69b5e7ddaa5)