Ninja build file parser
This adds support for parsing Ninja build files in support of go/agp-custom-cxx which is part of the project to allow opening of AGDE-driven build.gradle.

The code is factored into 4 layers for unittesting and separation of concerns. The main entry point is [streamNinjaBuildCommands] and the other layers are implementation details. So [streamNinjaBuildCommands] is probably a good place to start reviewing. If you prefer a bottom-up approach then [streamUnescapedNinja] is the right place to start.

Bug: 192006965
Test: Fully unittested along with some basic fuzz testing.

Change-Id: Ib7f45df0b7d39ce6aaecaaed3a442cb0116ec73a