Revert "Use testFixtures variant name for testFixtures resolved artifacts"
This reverts commit 9f2dc78c46f3b3a6c056178f9d1f869e34b12cd3.

Reason for revert: variantAttr is no longer used in v2 model, however
in v1, we still need to select the correct variant in case of
testFixtures module dependency. And since testFixtures is not
supported in v1 models, it's better to set the variantAttr to the
main variant instead of the testFixtures variant so the IDE could
pick up the main variant directly.

Change-Id: I0e2613b76ec2f6076201c8256d3c109923476322