add API to register new source types to the DSL and variant.
new source types cannot be added to the AndroidDirectorySourceSet
directly because it would allow users to add directories to just one
SourceSet like main or debug, which does not seem very useful.

Instead, users need to use an API on androidComponents to ensure
the new folders are added to all the source sets.

The added variants are available through the variant API using the
Sources interfaces. Variant specific folders can be added either
directly (non generated) or as a Task output (generated).

Test: Added gradle-recipes.
Bug: N/A

Change-Id: Iaf859060e8ad1f10f18211fed3eabc65dfca27ce