Revert "Add a sourceset field to LintModelModuleLibrary"
This reverts commit 5111dc704e61f401c4fe8e7c3516963a9a8568b1.

Reason for revert: Moving from using a sourceset field to rely on the
artifact address to distinguish between different sourceset artifacts.
There is no need currently for lint to know whether a library or a
module is a testFixtures component of a library or a module. By
moving to using the artifact address lint will deal with them as
separate artifacts.

Change-Id: Id2b5cdee0d13c36d15d53b85b9674329c6c6403b