Always output a lint failure
Since the lint result task can be up-to-date, there is now a separate
task to output the lint results to stdout/stderr as configured, as well
as being the task that fails if there are any errors in the analysed
code.

However, that currently means if text report is not enabled and there
are errors that they won't be any output about what is broken.

Add logic in the AndroidLintTextOutputTask to show the first error if
there are errors that will fail the build. Always output an intermediate
text report to support this.

Updated the unit test and integration test to verify this behavior.

Fixes: 205721223
Test: Included
Change-Id: Icf4305f43f6147312ebf6494d39424a46d4db79e