Fix detector options for per-file configurations and add test support
This CL fixes a bug in the lookup of per-file configurations (it should
be using context.findConfiguration, not getConfigurationForFile on the
configuration hierarchy).

It also adds test support for options, such that you can easily from a
unit test configure the test to run with particular values of the
option instead of having to manually configure a lint.xml file as a
test source.

It also adds validation of configured values for options that specify
a minimum or maximum range, as well as for basic data type errors (such
as making sure a boolean option specifies true or false, a numeric
option specifying a number, and so on.)

Finally it updates the docs to both describe the new testing support as
well as some notes on how to support options in versions older than 7.2
and how to use range constraints.

Test: Unit test included
Bug: Not filed
Change-Id: I306229f249abebd714909f781ae7b6c1c018bb47
(cherry picked from commit ca7f6d1fb338973032844a44daec8691ff1d176e)