Handle @Dimension values with different units
This CL cleans up several things related to handling of androidx
annotations.

First, it implements proper support for the unit attribute of the
@Dimension resource; before this it wasn't looking it up and would just
assume @Dimension == @Px.

Second, it fixes validation of the RequirePermission container
annotations; the attributes for @RequirePermission are not present on
the container Read and Write permissions and should therefore not be
enforced.

Finally, it fixes a few other annotations related issues, required by
the annotations inference code over on the IDE side: exposing access to
a few constants and methods, etc.

Test: Regression tests included
Bug: Not filed
Change-Id: I95d572934c6137efe5cadc8fbbae00f268af155e