Strip merged value resource files of comments
Comments in values resource files are not removed beyond one level of
nesting. This change removes comments from resources such as style.
There is still a risk of including comments for menu resources (as they
are capable of having deeper levels of nesting), however this resolves the
majority of cases.

Bug: 187865493
Test: MergeResourcesTest
Change-Id: I3cff56203b0baa3a320562b20fa9ebc0c7219144