Add --missing-baseline-is-empty-baseline flag to lint
Previously, lint would always write a baseline file if a baseline file
was specified but the file did not exist.

With this change, the --missing-baseline-is-empty-baseline flag will
cause lint to not write a baseline file when it does not exist.

If --missing-baseline-is-empty-baseline and --update-baseline are both
used when the specified baseline file does not exist, then a new
baseline file will be written iff there are lint issues.

Bug: 188666845
Test: LintBaselineTest
Change-Id: I2734d0c9107d6f4c87709972339a6f50fe854a1e