moved ASSETS from multiple artifacts to single.
multiple artifacts are not replaceable and it creates issues when
adding elements to ASSETS as it needs to be merged. Instead users should
use the [Sources] APIs to access assets sources and add custom
directories that way.

assets are now moved to be a single artifact type which mean that it
can be replaced if necessary.

Test: Adapte existing API tests.
Bug: 217891921
Change-Id: Id7c077860ce1649780f6fce551e178a1b0058dc1