removed PROJECT_CLASSES_DIRS/JARS public artifacts
Those artifacts were scheduled to replace the ALL_CLASSES versions
to better indicate that the scope was the module scope. Instead, it
was decided to use a dedicated API that will be available in Electric
Eel and up. Therefore reverted the changes to continue use the
ALL_CLASSES public artifact types until they become deprecated in
Electric Eel

Test: Adpated existing
Bug: N/A
Change-Id: I18bb265fccec3769a9fb655c964058692e1627f2