LE: Close channel on timeout
A command ran by the Installer can timeout on the device *after* the Installer Client timeout on the Host machine. When this happens, the installer command writes back an InstallerResponse which remains in
the OS socket buffer, dormant. This is what we call a "desync".

From there, a request get the response from the previous request (and the current response go to the OS socket buffer). To break this cycle, we close the channel on timeout which will trigger a clean installer respawn on next request.

Bug: 232146753
Test: Later (urgent fix)
Change-Id: If31e16548b079473a5e46dd7dd0d4a15efa06946
(cherry picked from commit 34f69a4956d1c1105dd0b0174939c96366f94b19)