adblib: shell command builder API
Unify all variations of `shellXxx` methods and extensions
to use a single underlying `shellCommand` API.

The `ShellCommand` implementation supports using the
most relevant protocol depending on the target device
capabilities, basically downgrading from `shellv2` to
`exec` or `shell` when `shellv2` is not supported.

The `ShellCommand` implementation also supports any
`Shell(V2)Collector`, setting various timeouts, and so on.

Note: There will be a follow up CL to clean up the various `shellXxx`
extensions, since some of them are now redundant, and some of the
existing ones should not need to maintain backward compatibility.

Bug: 188067646
Test: Existing + included
Change-Id: Iec2d0fa5bb0679fb0fd3eec63d465b10a97c5311