fakeadb: Switch to SocketChannel
Sockets from `java.net.Socket` don't support cancelling
pending I/O when threads are interrupted. This creates
issues when shutting down FakeAdbServer, specifically
if there are `JdwpCommandHandler` active, as their
implementation waits for the next `JdwpPacket` on
the socket.

With this change, fakeadbserver uses `ServerSocketChannel`
and `SocketChannel` (which are mostly compatible with the
older java.net equivalent) so that thread interruption, which
is required when shutting down a ThreadPoolExecutor,
is fully supported (i.e. any pending socket I/O operation
throws an exceptions when cancelled), as mentioned at https://docs.oracle.com/javase/7/docs/api/java/lang/Thread.html#interrupt()
"If this thread is blocked in an I/O operation upon an interruptible channel then the channel will be closed, the thread's interrupt status will be set, and the thread will receive a ClosedByInterruptException."

All in all, this ensures that
`FakeAdbServer.awaitServerTermination()` succeeds even when
there are active command handlers waiting on a socket I/O
(e.g. InputStream.read).

Bug: 237420199
Test: Existing (fakeadbserver is used in *many* unit tests)
Change-Id: I7d29cc49b7c9bd457818280d050a9669f3520b63