Stop supporting keyboard events in Event Timeline
For pre-O (API 25-), the instrumentation is done at compile
time. For O+, the instrumentation is done through JVMTI
at the beginning of runtime. Both approaches are updated.

We leave the door open to possibly enable the functionality
through a server-side flag for L - Q (API 21 - 30), in case
user feedback indicates a strong desire of it. If it
happens, we need to figure out a proper way to support it
for API 31+ as the implementation is out of sync of
platform.

Bug: 211154220
Test: Update existing tests as if the functionality is always enabled.
      Manually tested the flag because this change involes many
      parts of Studio including build, deploy, running and
      interacting with the app, and running profiler and the daemon.
Change-Id: Ib037bf98f65747fee8c6a32e18eeb2b4501a012b