Fix generated directories not added for resources
When using the new variant API addGeneratedSourceDirectory for
resources, the directory was actually not added to the merge
resource task and ended being ignored.

The code initializing the input FileCollection did not add
the new variant API resources directories, just the ones added
from the old variant API.

Added a functionality in our tests to provide a buildSrc sub
project. This is used in the regression test to add a custom
plugin using the new Variant API.

Test: Added a regression test
Bug: 237421684
Change-Id: I57240b55fa4ffcd72fa67ee4daa2a9aef15c1cda