Fix analytics service not initialized
This issue happens when we do a native model query for a composite build
where the dependency substitution is used. And the root cause is we make
decisions on initialize analytics service of the included build based on
the task names of the main build. The fix is to always make decision
based on the current build.

Bug: 226095015
Test: AnalyticsCompositeBuildTest
Change-Id: I905dd70a2e7ccf05b3a7da1c83e2af4260b71ece