Allow plugin-specific state classes
Currently, the Activating and Deactivating states have job and
timeoutTracker fields to facilitate timeout and cancellation of
activation and deactivation. However, in the current design, this is the
responsibility of the plugin, not the provisioner; thus, these are
internal implementation details that should not be exposed.

We can keep DeviceState sealed, and open its subclasses; this provides
a clean public interface to be used by the clients, while the plugins
can store private data in the subclasses if desired.

The Activating and Deactivating subclasses in the local emulator plugin
deliberately shadow the public class names to help ensure that they are
used rather than the public classes.

Bug: 241544067
Test: n/a
Change-Id: I9be8b301445659d8a918a1c991c80c3987471493