ddmlib: Make DebugViewDumpHandler more abstract
The DebugViewDumpHandler surface area was too large, specifically
exposed `ClientImpl` (an implementation detail), making it difficult
to implement the `Client` interface.

For example `listViewRoots(DebugViewDumpHandler callback)`
expects callers to implement a `DebugViewDumpHandler`,
but in practice, the only thing callers are expected to provide
is a Ddms Chunk Type that are interested in, and a callback
to invoke when the ddms packet reply payload is available
(as a ByteBuffer). So there is no need for callers to be
exposed to `ClientImpl` implementation details.

This change is required to be able to provide a
non-ddmlib implementation of `Client` via the
`ClientManager` extension point.

Bug: 241411829
Bug: 188067646
Test: Existing
Change-Id: Ia6e9bbb0eb5c85d4e72a48bf1fe4f29f94a6995d