change asFiles to return a list instead of singleton
The asFiles was returning a singleton as we assumed there could
only be one element in the list but that's not true if users
added directories. Changing the return type to List<> forced
the storage to contain mapped value rather than the actual
property contain the final producer.

This causes issues if the users wants the read the list during
configuration as Gradle will generate an exception when a
mapped value is resolved.

Removed a test that was resolving the sources during
configuration.

Test: Existing
Bug: 251761446

Change-Id: Icdd9e5de80d2714cf364b6e6df965579bd9df6b5