Remove unused UserOptions usages
The trace mananger and trace manager tests were
parameterized to use either UserOptions or the
new technology-specific options protos (as part
of the new unified TraceConfiguration proto). Now
that it has been verified that the new options are
working fine, the old/unreachable code can be
removed. There were other minor instances or uses
of the UserOptions proto that were removed.

Note: This change focused solely on the cpp side of
UserOptions usages. Any proto usages were not considered.

Fixes: 259745930
Test: updated & manual
Change-Id: I4a26b2d4dae31d6089c4832806f3f3d418e3955f