reintroduced lateAdditions for old variant API support
This reverts commit e67ce246d5ab0048190296981de8b52191e28913.
which itself was a revert of adding lateAdditions capabilities
to the sourcesets.

The lateAdditions are still useful but non functional without
the other changes introduced in this change.

The directories will be added into an existing DirectoryEntries
(identified from its sourceset name) so it is added with the right
priority.

Using a FileCollection to retrieve the list of directories was not
functional as task dependency was not properly carried. Instead,
use the Provider.zip method to flatter a List<List<>> into a
single list and provide the source directories as a
Provider<List<Directory>> which carries correctly task's
dependencies.

Finally, some Tasks that need to have access to the final list
of source directories at configuration time (which mean before
the old variant API has completed) can now register an action
to the CreationConfig. That action will be ran after the
variant API has completed and can be used to obtain the final
list of sources during configuration time.

Test: Added regression integration test.
Bug: N/A

Change-Id: Ia081409358123fd5a251fcd0801b8c5b7d2376c2