Add automatic binary jar file stubbing to lint's test infrastructure
When testing detectors which are accessing stubs of a particular
library, it's often useful to include those stubs as *binary* classes,
since there are differences between the PSI you get from a library jar
and from a local source file. As just a simple example, constants will
have been inlined.

Until now, creating binary stubs was cumbersome; you'd need to compile
them somewhere, and then copy in the gzip compressed, base64-encoded
class files.

This CL adds a new test DSL method which makes this simpler:

   mavenLibrary(artifact, *sourceFiles)

where you can specify the stubs as simple source files, and lint will
dynamically create class files on the fly, package these as a jar, and
place the jar in the test project instead.

The way this works is that lint will first parse the source file using
UAST, and then it visits the UAST and uses an ASM ClassWriter to write
a series of corresponding class files.

This CL also updates some existing tests which used base64gzipped stubs
over to using this new mechanism (but the real driver for this feature
is an upcoming detector for dependency analysis which will needs to test
a lot of library stubs.)

As part of this I discovered and fixed two bugs:

- The code to generate a method descriptor produced an incorrect
  descriptor when you call a constructor of a non-static innerclass.  In
  this case, the bytecode will have a synthetic argument (to the
  outerclass) as the first parameter, which was missing/put in the wrong
  place.

- The test mode support for testing source-only and bytecode-only
  versions wasn't working right for compiled test files packaged as jars
  (since the sources were placed inside the .jar file).

Test: Existing tests, and many new unit tests as well
Bug: Not filed
Change-Id: Ie644d6b8065205980780af198ae25469c35eca62