Replace MemoryNativeSampleData with MemoryTraceInfo
A new proto message (that duplicates that of CpuTraceInfo)
has been created named MemoryTraceInfo. A corresponding
replacement of respective event kind was made to replace
MEMORY_NATIVE_SAMPLE_CAPTURE, called MEM_TRACE, similar
to the CPU_TRACE event.

This change faciliates the conversion of native heap sampling
to use the trace manager used by cpu tracing. In future changes,
while the kind of event needs to be distinct, we can look to
merge the CpuTraceInfo and MemoryTraceInfo into one TraceInfo.

Bug: 263172101
Test: updated
Change-Id: I5378a0c5260166e8998caa401ae8b2717137d04c