Add controls for the IDE to decide whether to build the runtimeclasspath
This change adds a new method to the ModelBuilderParameter allowing
the IDE to prevent the model builder from building the runtime classpath.

Since AGP can't be opened with an older version of studio there should
be no case where this paremter exists but is not set. The reverse will
also be safe since reguardless to whether this is set the ModelBuilder
won't read it and will always produce the full classpath. This is
handled by Gradle and its use of proxys.

Bug: 267066187
Test: added test model snapshots
Change-Id: I5479d25e500510378093fafc50f6bc96be1a8cf3