changed Svg2Vector user to new API
The unused VdCommandLineTool was not updated as part of the
Change-Id: Ib334788a63c0b6ae7bcf61e721ae0c90312fa50f

when an API was moved from using File to Path.
This failed to build in Intellij even if the file is unused.
I would support deleting the file altogether but since I am
not familiar with it, I chose to update it.

Test: manually build tools/base in IntelliJ
Bug: N/A
Change-Id: I268356cf23d82e5c12c7b373f6f4e9b8310dcd7b