adblib: Improve behavior for WAIT packet
A previous change (I399ec064ab084ff4da1ba9cc07336515e0813522)
was made to keep the JDWP session open for a as long
as the timeout used by the process properties collector
(15 seconds currently).

However, given that close a JDWP session in the `WAIT` state
leaves the session completely unusable until the process is
terminated, this change updates the behavior to keep the
JDWP session open until either 1) the process is terminated
or 2) an external debugger (e.g. Android Studio) attaches
to the process. There is no timeout involved.

Bug: 271466829
Test: Included
Change-Id: I290cac4a107553f395e58a031ed2ecf03b8ee209