Make TestAnalytics callable from WorkerAction
Make recordOKInstrumentatedTestRun and recordFailedInstrumentatedTestRun
callable from inside a worker action.

This came about as ArtifactCollections are not serializable and
therefore cannot be included as parameters into the worker action. As
such we cannot call the existing functions. The proto class, however is
serializable so if we generate the proto before starting the worker
action we can call the new interface just fine.

Bug: 249817356
Test: Manual + eventually the TestLabPlugin Instrumentation tests
Change-Id: I40b70027ce69e5c3c760c97d344e8b478e2a19ad