Simplify KnownVersionStability internals a little
We no longer need to round-trip version numbers or parse into multiple
different kinds of version.  Instead simply define stability on a
given Component and propagate that through.

Leave the old version around because who knows who's used it.  (It's
also used as a precondition checker in a couple of other tests,
harmlessly.)

Bug: 242691473
Test: existing tests adjusted
Change-Id: I27dab3fc20004eacd979f03bd8c7ee7fc542c721