Make desugaring graph used for dexing transforms relocatable
To support incremental dexing transforms, we use a desugaring graph
which contains the class files' paths and their dependencies.

Previously, the file paths were absolute, which means they are not
relocatable and could break incremental dexing after a cache hit
(e.g., bug 266599585).

With this commit, we store only relative paths in the desugaring graph.
This requires that all files in the graph must share one single root
directory so that they can be converted to relative paths.

This then means that the desugaring graph/incremental dexing transforms
is currently only possible for inputs that are directories (not jars)
and for minSdk >= 24 (where dexing/desugaring doesn't require a
classpath). However, this was already the case before for a different
reason (see comment in `DexingArtifactConfiguration#registerTransform`),
so this is not a regression.

Bug: 266599585
Test: Added CacheableDexingTransformTest
Change-Id: Icd6cdd698baf6d64b4bfa3c4adf896463fc218f3