Set default power data processing query
Before, the way in which power data was processed
was determined by different SQL queries in the trace
processor daemon. The query itself was determined
by the display mode feature flag for the power profiler.
However, because data processing is now going to be done
on studio-side, we can use the same SQL query for all
display modes. This shared query being essentially the
raw data.

Thus, this change removes the display mode field
from the PowerCounterTracksParameters proto. It also
sets a default SQL query for power data in the trace
process daemon.

Bug: 254116554
Test: updated
Change-Id: I014dca2ac5238fcea466948a47a70bce24aa2d34