[FTL GMD] support shared file uploads
Added support to the testlab storage manager to support the upload and
use of shared test files accross device tasks. This only implements the
ability to upload shared files in the StorageManager class, and the
functionality will be added in a later CL.

The fundamental idea is that the file can be uploaded from multiple
different computers in multiple different states. So the Testlab plugin
needs the ability to not only have uploaded multiple different versions
of a given file, but also the ability to differentiate between them.
This is done by prepending the sha256 hash to the file name when it is
uploaded to storage.

In addition, to ensure that the testlab plugin doesn't unnecessarily
rehash large files, the FileHashCache was created to keep track of
files' hashes by caching on both the file name and the last modified
timestamp. Id est, rehash only if it is a new file, or if a given file
has been changed.

Bug: 276517167
Test: Added new tests
Change-Id: Ia59948c2940f6491598a1897e7deef958da36d75