Support running lint on non-Android KMP Projects
This change also adds support for running lint on a KMP project with
android and jvm targets.

With this change, android lint tasks are not registered unless the
standalone lint plugin is applied.

TODO:
1. Support running lint on multiple jvm targets
2. Initialize the values of the android-specific lint DSL using the
   global lint DSL
3. Have a single lint report task (and a single lint report file) per
   KMP module. With this change, there is a lint report task per target
   for KMP modules.

Some cleanup with this change:
Most lint tasks no longer have their group set to VERIFICATION_GROUP
because that caused too many lint tasks to be listed when a user would
run ./gradlew tasks.

This change also increases the shard_count from 6 to 8 for the AGP lint
integration tests. The bazel target was already timing out frequently
on studio-win, and this change adds more integration tests.

Bug: 196838286
Test: KotlinMultiplatformAndroidLintTest
Change-Id: Ib32f22c0e2b921d4c21f9630ef4916ed98785556