[lint] Unify the LintModel representation of non-Maven artifacts
Actually that is a bit too much of a claim.  What I am fairly sure is
true, though, is that LOCAL_AARS is not special enough to be the only
thing tested for in the routine that generates a name for a lint
diagnostic string; we have other prefixes that probably should be
treated as internal (__local_asars__, __wrapped_aars__, artifacts,
...).

Sync now provides the IDE with information regarding whether the
library has valid maven/Gradle coordinates, which means that the IDE
can name the library accordingly, with a special NON_MAVEN tag for
names that are non-maven.  I don't really know how the Lint cli builds
its models nor how to intercede, but the only user of this is to
generate a lint diagnostic string, so if this does change behavior it
might not be too harmful.

Bug: 279886738
Test: existing
Change-Id: Ifcf4b331cd4082931edd2c662107ed34772a2fe1