Use const_string_jumbo over const_string
It turns out the DEX code ART provides us during class tranformation has a
string pool identical to the original classes.dex. It is not splitted
per class like a typical DEX splitter.

That means it is possible for us to run into the 65k limit if we are
adding a few strings to the pool.

The solution is just use const_string_jumbo all the time.

We can protentially improve the slicer library to use the smallest
instruction possible but that might require multiple pass because
strings need to be sorted.

As we are just sending the dex code right back to ART in memory
and not over the network, I believe the overhead is minimal.

Fixes: 295079565
Test: Existing Tests should cover all instrumentation code.
Change-Id: Ie87ebf0030b7563bbb2fe163db1837c46d5b6014
(cherry picked from commit e0b5fe100ab8efe7fb4e424ac32252c5fae6abfa)