Refactor dexing transform artifact attributes
When setting up dexing transforms (or any transforms in general), we
need to make sure variant-specific parameters are captured as
attributes; otherwise artifact consumption could be incorrect (e.g.,
we had bug 144668341 in the past, and bug 293206305 recently).

This commit ensures all variant-specific parameters are captured as
attributes *automatically* (using `toString()` of a `data class`) so
that we don't forget to update the attributes when the parameters
have changed.

Bug: 293206305 (this commit does not yet fix this bug, it only refactors
     the code to set up a pattern to use later)
Test: Existing tests (refactoring change)
Change-Id: I46f0ea6707536a2f00974fbc4fd1134e02579e97