Make Anonymizer return null if salt is not present
Currently, if Anonymizer is used before AnalyticsSettings.initialize is
called, an empty byte array is used as salt. Instead, we should abort.

Remove the unused logger parameter. Remove utf8 from the name, since how
we encode the string before hashing it is an implementation detail, and
remove the unused variant that allows passing an alternative Charset.
IOException is never thrown, so remove the throws clause and adjust
callers to handle null instead.

Bug: 297544591
Test: updated / added
Change-Id: Ic3139cea98e31a19731777dd0a2037fdbb7efe20