Provide BooleanOption to run dexing transforms with full classpath
The new BooleanOption `android.useFullClasspathForDexingTransform` when
enabled means that dexing transforms will use the full classpath (if
desugaring requires a classpath). This classpath consists of all
external artifacts in addition to the input artifact's dependencies
provided by Gradle.

This option is useful when some dependencies are missing from the input
artifact's dependencies (see bug 230454566), so the full classpath is
needed.

If dexing transforms do not require a classpath, this option is not
used.

Bug: 230454566
Test: New DesugarCompileOnlyDependencyTest
Change-Id: Ibee74d95f42e58f45ab873724ec6d1a59f7e1f37