Use the classes jar variant to match our library representation with KGP
As we're using the AAR_OR_JAR artifact to build library dependency
representation, the attributes of that artifact might be different from
the attributes coming from the kotlin gradle plugin resolution as we're
using CLASSES_JAR to resolve the configurations. This change uses the
coordinates coming from the CLASSES_JAR resolution as a basis of
matching our library representation with the kotlin ones so that we can
attach the models correctly.

Also since Gradle might resolve a dependency to two different artifacts
pointing to the same file but different in the attributes due to an
issue in Gradle, this change relies on the coordinates of the original
artifact instead of our library representation since we dedup this case
but the kotlin plugin doesn't.

Fixes: 301431703
Test: existing
Change-Id: I6e2c1e563c8e515254fa533f2ff377a494bdf4a5