Use main split when generating test config
When generating test config, we obtain the manifest path by loading
merged manifest directory through BuiltArtifactsLoader. In case of abi
splits, multiple manifest files are generated(one per abi). We borrow
the the concept of main split determination from VariantOutputList by
choosing the builtArtifact that closely matches with the target device
configuration. In this change, we are reusing the logic of finding the
the buit artifact that best match the target configuration.

Bug: 293547829
Test: GenerateTestConfigTest
Change-Id: I69bc7c15857b3c4c402afb58f9183558d768176f