Add Kotlin scope function support to PendingIntentMutableImplicitDetector
This change adds support for Kotlin scope functions such as apply, also,
etc, which require extra care with a separate DataFlowAnalyzer. Since
most of the functionality of that analyzer is the same as the existing
IntentSimpleNameAnalyzer, this change creates a parent class
IntentDataFlowAnalyzer that evaluates whether a single intent is
implicit. The current IntentSimpleNameAnalyzer now extends that new class
and still has support for Java arrays.

In terms of the scope functions, "with" is a method call, so it is
handled by getIntentTypeForCall(). The rest of the scope functions are
selectors of UQualifiedReferenceExpressions, so they are handled by
getIntentTypeForQualified().

Bug: 274913202
Test: PendingIntentMutableFlagDetectorTest
Change-Id: I1406654c945993b007d3c33ce5b1a5c35355a3fa