Do not run dexing task on subprojects' classes unnecessarily
Classes should be dex'd by either the dexing task or the dexing
transform, not both.

We accidentally changed this behavior for subprojects' classes in a
refactoring (commit 14a7b76 - AGP 8.0), meaning that
DexArchiveBuilderTask would dex subprojects' classes even though they
are to be dex'd through artifact transforms.

This commit will revert that accidental change.

PERFORMANCE: The performance regression was probably large because we
were dexing *all* subprojects' classes *twice* (see bug 309101832 for
some numbers). This commit fixes the regression, so the performance
improvement will be of the same scale.

Test: Updated DexingArtifactTransformMultiModuleTest
Bug: 309101832
Change-Id: I323e680b0f8bec5c274fb82e7ff44ea4cb02880f