Attach task dependencies for default Proguard files
Default Proguard files are generated by a task, so it should have a
dependency on that task.

In previous commit bd6b069, we fixed the missing task dependency for
ExportConsumerProguardFilesTask. However, there may be other usages of
`OptimizationCreationConfig.getConsumerProguardFiles` which still don't
have task dependencies because the return type is `List<File>`.

In this commit, we will change the type of
`OptimizationCreationConfig.consumerProguardFiles` from `List<File>` to
`Provider<List<RegularFile>>` (with task dependencies), so that all of
its usages will have task dependencies.

Test: DefaultProguardFilesTest
Bug: 295666695
Change-Id: I3f8cb3245dab0ca7f0eb4e63694efc1b17d810b4