Fix integer max value parsing and DM range issue
Recently, the DM range was updated so that the Integer.MAX_VALUE was
used to signify that there is no upper bound for API level that a DM
file can be installed. This needed changes in AGP to parse the max value
correctly, and also in the InstallVariantTask to check for any value
greater than the highest API level range.

Bug: 309851946
Test: existing
Change-Id: Ia85359c234499eb1b0436c03483d9de0a571a9b6