Validate lint test inputs
This CL adds some basic validation of some of lint's input
files -- both source inputs and network inputs. I was debugging
why current-version lookup for the compose-bom wasn't working
from a unit test, when I discovered that the corresponding XML
file accidentally didn't have an `.indented()` call on it, and
because this was an XML file with a prologue, that makes the XML
invalid (you cannot have whitespace before <?xml version=...?>
so the XML parser threw an exception). I added some checks and
discovered several other places where the XML was invalid. This
CL adds some additional validation for other network mock data,
such as json as well.

Test: Unit tests included
Bug: Not filed
Change-Id: Ic074214e04e50ce5eeff5991455b14f0d9602104