Ensure ExtractAarTransform is deterministic
ExtractAarTransform extracts contents of an AAR to a directory. If the
AAR doesn't have classes.jar, it will create an empty classes.jar *with
a manifest* inside the jar.

The issue is that the manifest entry has the current timestamp by
default, therefore the generated classes.jar will not be deterministic.

To fix this issue, we could:
  1. Ensure that the manifest entry has a fixed timestamp
  2. Or just don't create the manifest entry at all because it's not
     needed (commit bb53cc9a9 which introduced this logic did not
     mention why it was needed)

We'll go with option 2 in this commit because it both fixes the issue
and simplifies the code.

Bug: 315336689
Test: ExtractAarTransformTest
Change-Id: I94ac56ef8275785c89741024173961e631f19b5c